### 2019-11-07 #10 @c-torre
Review Request: C. de la Torre-Ortiz and A. Nioche
Dear ReScience editors,

I request a review for the following replication:

**Original article:** https://doi.org/10.3389/fncom.2015.00149

**Scientific domain:** Computational Neuroscience
**Suggested editor:** @oliviaguest
