### 2019-11-08 #10 @anne-urai
I don't really have expertise in neural networks, but my colleague @jamesproach is, and is happy to do this. He'll register as a reviewer first.
