### 2019-11-21 #10 @bobaseb
The authors do a good job in transforming the original model based on individual neurons to a population model. New parameters are introduced to this effect and the original model's dynamics (i.e., retrieving a memory while inhibiting others under periodic inhibition dynamics) is broadly reproduced.

The code is clear and human-readable with enough comments to get the gist of each computation. I was able to run it with ease on a linux machine.

Three specific questions came to mind when reviewing this work.

1. Why is the sparsity parameter different? Shouldn't it be 0.1?

2. Where is the recall activation threshold parameter defined?

3. Why is there no attempt to plot something similar to figures 3 to 6?

Since the original model has been modified and only the first figure of the original paper seems to have been replicated, perhaps this is best suited as a partial replication.
