### 2019-11-26 #10 @jamesproach
The authors reproduce simulation results from Recanatesi et al using a model of associative memory retrieval which accounts for the sequential activation of memory though two network features: (1) synaptic connections which enforce sequential associations between particular memories and (2) periodic inhibition. This is an extension of the Hopfield Network model of auto-associative memory. In their reproduction, the authors employ a dimensionality reduction proposed in the original article which groups units that present the same activity across all memories reducing the size of the dynamical system by several orders of magnitude.

The authors were able to reproduce the simulation results from the original article with slight modifications to the parameters. The associated Python code is clearly written, well commented and fully reproduces the figures in the manuscript.

Major Comment:
What accounts for the difference between the results with the original and reproduction variable? As a work aimed at reproduction, presenting differences between dynamics with the original parameters and the scaled reproduction variables would be useful. In some cases parameters need to be changed by N (10^5). Presentation of the differences would be helpful to understand this.

Minor Comments:
After running the network for several different seeds, I found that there were several periods where the same memory was recalled in sequence. The original article presents a simulation with up to three repetitions while the provided code often produces repetitions of 6 or longer, even 14 in one case. Is this a meaningful difference resulting from the difference in parameters?

Could the authors provide a reproduction of the simulation results in figures 3-6.
