+++ main.py
@@ -20,9 +20,13 @@ import os
 
 import numpy as np
 from tqdm import tqdm
+from deco import concurrent, synchronized
 
 import tools.plots as plot
 import tools.sine_wave
+
+from multiprocessing.pool import Pool
+from multiprocessing import cpu_count
 
 FIG_DIR = "fig"
 os.makedirs(FIG_DIR, exist_ok=True)
@@ -30,6 +34,7 @@ os.makedirs(FIG_DIR, exist_ok=True)
 np.seterr(all="raise")
 
 np.random.seed(123)
+
 
 # === Parameters ===
 # Architecture
@@ -42,7 +47,7 @@ threshold = 0
 gain_exp = 2 / 5
 # Hebbian rule
 excitation = 13000
-sparsity = 0.01
+sparsity = 0.1
 # Inhibition
 sin_min = 0.7
 sin_max = 1.06
@@ -65,177 +70,261 @@ param_current = 4.75
 # ==================
 
 
-print("<[Re] Recanatesi (2015)>  Copyright (C) <2019>\n"
-      "<de la Torre-Ortiz C, Nioche A>\n"
-      "This program comes with ABSOLUTELY NO WARRANTY.\n"
-      "This is free software, and you are welcome to redistribute it\n"
-      "under certain conditions; see the 'LICENSE' file for details.\n")
+def inside_loop_connectivity_matrices(arg):
 
-# Compute memory patterns
-print("Computing memory patterns and neuron populations...")
-memory_patterns = np.random.choice([0, 1], p=[1 - sparsity, sparsity],
-                                   size=(num_neurons, num_memories))
+    i, num_pops, pop, backward_cont, forward_cont = arg
 
-# Compute populations
-pop, neurons_per_pop = np.unique([tuple(i) for i in memory_patterns],
-                                 axis=0, return_counts=True)
+    regular = np.zeros(num_pops)
+    forward = np.zeros(num_pops)
+    backward = np.zeros(num_pops)
 
-num_pops = len(pop)
-
-neurons_encoding = [(pop[:, mu] == 1).nonzero()[0] for mu in
-                    range(num_memories)]
-
-# === Other pre-computations ===
-num_iter = int(t_tot / t_step)
-relative_excitation = excitation / num_neurons
-time_param = t_step / t_decay
-
-# Inhibition
-sine_wave = np.zeros(num_iter)
-
-# Connectivity
-forward_cont = np.arange(num_memories - 1)
-backward_cont = np.arange(1, num_memories)
-
-# Noise
-noise = np.zeros((num_pops, num_iter))
-
-# Neuron dynamics
-firing_rates = np.zeros([num_pops])
-current = np.zeros([num_pops])
-average_firing_rates_per_memory = np.zeros((num_memories, num_iter))
-currents = np.zeros((num_pops, num_iter))
-currents_memory = np.zeros((num_memories, num_iter))
-# ==============================
-
-# Compute sine wave
-print("Calculating inhibition values...")
-
-for t in range(num_iter):
-    sine_wave[t] = tools.sine_wave.sinusoid(
-        min_=sin_min,
-        max_=sin_max,
-        period=t_oscillation,
-        t=t,
-        phase_shift=phase_shift * t_oscillation,
-        dt=t_step
-    )
-
-inhibition = -sine_wave
-
-# Compute weights
-print("Computing regular, forward and backward connectivity...")
-
-regular_connectivity = np.zeros((num_pops, num_pops))
-forward_connectivity = np.zeros((num_pops, num_pops))
-backward_connectivity = np.zeros((num_pops, num_pops))
-
-for i in tqdm(range(num_pops)):
     for j in range(num_pops):
[…]
