+++ main.py
@@ -17,6 +17,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
 import os
+import pickle
 from multiprocessing import cpu_count
 from multiprocessing.pool import Pool
 
@@ -54,7 +55,7 @@ phase_shift = 0.75
 cont_forth = 1500
 cont_back = 400
 # Time
-t_tot = 2  # 14
+t_tot = 1  # 14
 t_step = 0.001
 # Noise
 noise_var = 65
@@ -62,8 +63,9 @@ noise_var = 65
 init_rate = 1
 first_memory = 7
 # Replication parameters
-param_noise = 1 #10
-param_current = 1 # 4.75
+param_noise = 0.01
+param_current = 1  # 4.75
+
 
 # ==================
 
@@ -159,167 +161,179 @@ def compute_connectivity_matrices(num_pops, pop, forward_cont,
 #     return currents
 
 
-def main():
-    print("<[Re] Recanatesi (2015)>  Copyright (C) <2019>\n"
-          "<de la Torre-Ortiz C, Nioche A>\n"
-          "This program comes with ABSOLUTELY NO WARRANTY.\n"
-          "This is free software, and you are welcome to redistribute it\n"
-          "under certain conditions; see the 'LICENSE' file for details.\n")
+print("<[Re] Recanatesi (2015)>  Copyright (C) <2019>\n"
+      "<de la Torre-Ortiz C, Nioche A>\n"
+      "This program comes with ABSOLUTELY NO WARRANTY.\n"
+      "This is free software, and you are welcome to redistribute it\n"
+      "under certain conditions; see the 'LICENSE' file for details.\n")
 
-    # Compute memory patterns
-    print("Computing memory patterns and neuron populations...")
-    memory_patterns = np.random.choice([0, 1], p=[1 - sparsity, sparsity],
-                                       size=(num_neurons, num_memories))
+# Compute memory patterns
+print("Computing memory patterns and neuron populations...")
+memory_patterns = np.random.choice([0, 1], p=[1 - sparsity, sparsity],
+                                   size=(num_neurons, num_memories))
 
-    # Compute populations
-    pop, neurons_per_pop = np.unique([tuple(i) for i in memory_patterns],
-                                     axis=0, return_counts=True)
+# Compute populations
+pop, neurons_per_pop = np.unique([tuple(i) for i in memory_patterns], axis=0,
+                                 return_counts=True)
 
-    num_pops = len(pop)
+num_pops = len(pop)
 
-    neurons_encoding = [(pop[:, mu] == 1).nonzero()[0] for mu in
-                        range(num_memories)]
+neurons_encoding = [(pop[:, mu] == 1).nonzero()[0] for mu in
+                    range(num_memories)]
 
-    # === Other pre-computations ===
-    num_iter = int(t_tot / t_step)
-    relative_excitation = excitation / num_neurons
-    time_param = t_step / t_decay
+# === Other pre-computations ===
+num_iter = int(t_tot / t_step)
+relative_excitation = excitation / num_neurons
+time_param = t_step / t_decay
 
-    # Inhibition
-    sine_wave = np.zeros(num_iter)
+# Inhibition
+sine_wave = np.zeros(num_iter)
 
-    # Connectivity
-    forward_cont = np.arange(num_memories - 1)
-    backward_cont = np.arange(1, num_memories)
+# Connectivity
+forward_cont = np.arange(num_memories - 1)
+backward_cont = np.arange(1, num_memories)
 
-    # Noise
-    noise = np.zeros((num_pops, num_iter))
+# Noise
+noise = np.zeros((num_pops, num_iter))
 
-    # Neuron dynamics
[…]
