+++ main.py
@@ -55,7 +55,7 @@ phase_shift = 0.75
 cont_forth = 1500
 cont_back = 400
 # Time
-t_tot = 1  # 14
+t_tot = 1.0  # 14
 t_step = 0.001
 # Noise
 noise_var = 65
@@ -120,223 +120,185 @@ def compute_connectivity_matrices(num_pops, pop, forward_cont,
 
     return regular_connectivity, forward_connectivity, backward_connectivity
 
+def main():
+    print("<[Re] Recanatesi (2015)>  Copyright (C) <2019>\n"
+          "<de la Torre-Ortiz C, Nioche A>\n"
+          "This program comes with ABSOLUTELY NO WARRANTY.\n"
+          "This is free software, and you are welcome to redistribute it\n"
+          "under certain conditions; see the 'LICENSE' file for details.\n")
 
-# def inside_loop_currents(arg):
-#     weights_without_inhibition_pop, \
-#     inhibition_t, neurons_per_pop, firing_rates, \
-#     time_param, current_pop, noise_pop_t = arg
-#
-#     weights = (weights_without_inhibition_pop[:]
-#                + inhibition_t) / num_neurons
-#
-#     # Compute input
-#     input_v = np.sum(weights[:] * neurons_per_pop[:] *
-#                      firing_rates[:])
-#
-#     current_pop += time_param * (
-#             -current_pop + input_v
-#             + noise_pop_t)
-#
-#
-# def compute_currents(weights_without_inhibition, inhibition_t,
-#                      neurons_per_pop, firing_rates, num_pops,
-#                      time_param, noise_t, currents):
-#
-#     args = [
-#         (weights_without_inhibition[i],
-#          inhibition_t,
-#          neurons_per_pop,
-#          firing_rates,
-#          time_param,
-#          currents[i],
-#          noise_t[i]
-#          ) for i in range(num_pops)
-#     ]
-#
-#     with Pool(cpu_count()) as pool:
-#         currents = np.asarray(pool.map(inside_loop_currents, args))
-#         pool.close()
-#         pool.join()
-#
-#     return currents
+    # Compute memory patterns
+    print("Computing memory patterns and neuron populations...")
+    memory_patterns = np.random.choice([0, 1], p=[1 - sparsity, sparsity],
+                                       size=(num_neurons, num_memories))
 
+    # Compute populations
+    pop, neurons_per_pop = np.unique([tuple(i) for i in memory_patterns], axis=0,
+                                     return_counts=True)
 
-print("<[Re] Recanatesi (2015)>  Copyright (C) <2019>\n"
-      "<de la Torre-Ortiz C, Nioche A>\n"
-      "This program comes with ABSOLUTELY NO WARRANTY.\n"
-      "This is free software, and you are welcome to redistribute it\n"
-      "under certain conditions; see the 'LICENSE' file for details.\n")
+    num_pops = len(pop)
 
-# Compute memory patterns
-print("Computing memory patterns and neuron populations...")
-memory_patterns = np.random.choice([0, 1], p=[1 - sparsity, sparsity],
-                                   size=(num_neurons, num_memories))
+    neurons_encoding = [(pop[:, mu] == 1).nonzero()[0] for mu in
+                        range(num_memories)]
 
-# Compute populations
-pop, neurons_per_pop = np.unique([tuple(i) for i in memory_patterns], axis=0,
-                                 return_counts=True)
+    # === Other pre-computations ===
+    num_iter = int(t_tot / t_step)
+    relative_excitation = excitation / num_neurons
+    time_param = t_step / t_decay
 
-num_pops = len(pop)
+    # Inhibition
+    sine_wave = np.zeros(num_iter)
 
-neurons_encoding = [(pop[:, mu] == 1).nonzero()[0] for mu in
[…]
