### 2020-01-21 7a649e1
adding figure 3, linting, formatting
