+++ main.py
@@ -37,12 +37,12 @@ np.random.seed(123)
 
 # === Parameters ===
 # Architecture
-num_neurons = 10 ** 5
-num_memories = 16
+NUM_NEURONS = 10 ** 5
+NUM_MEMORIES = 16
 # Activation
-t_decay = 0.01
+T_DECAY = 0.01
 # Gain
-gain_threshold = 0
+GAIN_THRESHOLD = 0
 gain_exp = 2 / 5
 # Hebbian rule
 excitation = 13_000
@@ -66,7 +66,7 @@ first_memory = 7
 # Recall
 recall_threshold = 15
 # Replication parameters
-param_noise = 1000
+param_noise = 300
 # 10 we see saw peaks and bubbles and good transitions
 # print(param_noise)
 param_inhibition = 1
@@ -85,37 +85,25 @@ def inside_loop_connectivity_matrices(arg):
     backward = np.zeros(num_pops)
 
     for j in range(num_pops):
-        regular[j] = np.sum(
-            (pop[i, :] - sparsity)
-            * (pop[j, :] - sparsity)
-        )
+        regular[j] = np.sum((pop[i, :] - sparsity) * (pop[j, :] - sparsity))
 
-        forward[j] = np.sum(
-            pop[i, forward_cont] *
-            pop[j, forward_cont + 1]
-        )
+        forward[j] = np.sum(pop[i, forward_cont] * pop[j, forward_cont + 1])
 
-        backward[j] = np.sum(
-            pop[i, backward_cont] *
-            pop[j, backward_cont - 1]
-        )
+        backward[j] = np.sum(pop[i, backward_cont] * pop[j, backward_cont - 1])
 
     return regular, forward, backward
 
-
-def compute_connectivity_matrices(num_pops, pop, forward_cont,
-                                  backward_cont):
+def compute_connectivity_matrices(num_pops, pop, forward_cont, backward_cont):
     regular_connectivity = np.zeros((num_pops, num_pops))
     forward_connectivity = np.zeros((num_pops, num_pops))
     backward_connectivity = np.zeros((num_pops, num_pops))
 
-    args = [(i, num_pops, pop, backward_cont, forward_cont)
-            for i in range(num_pops)]
+    args = [(i, num_pops, pop, backward_cont, forward_cont) for i in range(num_pops)]
 
     with Pool(cpu_count()) as pool:
         r = list(
-            tqdm(pool.imap(inside_loop_connectivity_matrices, args),
-                 total=len(args)))
+            tqdm(pool.imap(inside_loop_connectivity_matrices, args), total=len(args))
+        )
         pool.close()
         pool.join()
 
@@ -129,49 +117,56 @@ def compute_connectivity_matrices(num_pops, pop, forward_cont,
 
 
 # xxx to remove
-class Tester:
-    pass
+# class Tester:
+#     pass
 
 
-tester = Tester()
+# tester = Tester()
 
 
 # xxx
 
 
 def main():
-    print("<[Re] Recanatesi (2015)>  Copyright (C) <2019>\n"
-          "<de la Torre-Ortiz C, Nioche A>\n"
-          "This program comes with ABSOLUTELY NO WARRANTY.\n"
-          "This is free software, and you are welcome to redistribute it\n"
-          "under certain conditions; see the 'LICENSE' file for details.\n")
+    """ Licensing and main execution """
+
+    print(
+        "<[Re] Recanatesi (2015)>  Copyright (C) <2019>\n"
+        "<de la Torre-Ortiz C, Nioche A>\n"
+        "This program comes with ABSOLUTELY NO WARRANTY.\n"
+        "This is free software, and you are welcome to redistribute it\n"
+        "under certain conditions; see the 'LICENSE' file for details.\n"
+    )
 
[…]
