### 2020-01-24 f82f9b1
formatted params and added seeds for all trials
