+++ main.py
@@ -25,15 +25,15 @@ import numpy as np
 from tqdm import tqdm
 
 import tools.plots as plot
-import tools.sine_wave
 import tools.recall_performance
+import tools.sine_wave
 
 FIG_DIR = "fig"
+
 os.makedirs(FIG_DIR, exist_ok=True)
 
 np.seterr(all="raise")
 
-np.random.seed(123)
 
 # === Parameters ===
 # Architecture
@@ -43,41 +43,42 @@ NUM_MEMORIES = 16
 T_DECAY = 0.01
 # Gain
 GAIN_THRESHOLD = 0
-gain_exp = 2 / 5
+GAIN_EXP = 2 / 5
 # Hebbian rule
-excitation = 13_000
-sparsity = 0.1
+EXCITATION = 13_000
+SPARSITY = 0.1
 # Inhibition
-sin_min = 0.7
-sin_max = 1.06
-t_oscillation = 1
-phase_shift = 0.75
+SIN_MIN = 0.7
+SIN_MAX = 1.06
+T_OSCILLATION = 1
+PHASE_SHIFT = 0.75
 # Short term association
-cont_forth = 1500
-cont_back = 400
+CONT_FORTH = 1500
+CONT_BACK = 400
 # Time
-t_tot = 14.0  # 14
-t_step = 0.001
+T_CYCLES = 14.0  # 14
+T_STEP = 0.001
 # Noise
-noise_var = 65
+NOISE_VAR = 65
 # Initialization
-init_rate = 1
-first_memory = 7
+INIT_RATE = 1
+FIRST_MEMORY = 7
 # Recall
-recall_threshold = 15
+RECALL_THRESHOLD = 15
 # Replication parameters
-param_noise = 300
+PARAM_NOISE = 700
 # 10 we see saw peaks and bubbles and good transitions
 # print(param_noise)
-param_inhibition = 1
+PARAM_INHIBITION = 1
 # print(param_inhibition)
-param_current = 1  # 4.75
-
+PARAM_CURRENT = 1  # 4.75
 
 # ==================
 
 
 def inside_loop_connectivity_matrices(arg):
+    """ Compute the loop inside connectivities to parallelize """
+
     i, num_pops, pop, backward_cont, forward_cont = arg
 
     regular = np.zeros(num_pops)
@@ -85,7 +86,7 @@ def inside_loop_connectivity_matrices(arg):
     backward = np.zeros(num_pops)
 
     for j in range(num_pops):
-        regular[j] = np.sum((pop[i, :] - sparsity) * (pop[j, :] - sparsity))
+        regular[j] = np.sum((pop[i, :] - SPARSITY) * (pop[j, :] - SPARSITY))
 
         forward[j] = np.sum(pop[i, forward_cont] * pop[j, forward_cont + 1])
 
@@ -93,7 +94,10 @@ def inside_loop_connectivity_matrices(arg):
 
     return regular, forward, backward
 
+
 def compute_connectivity_matrices(num_pops, pop, forward_cont, backward_cont):
+    """ Compute matrices with parallelization """
+
     regular_connectivity = np.zeros((num_pops, num_pops))
     forward_connectivity = np.zeros((num_pops, num_pops))
     backward_connectivity = np.zeros((num_pops, num_pops))
@@ -101,30 +105,19 @@ def compute_connectivity_matrices(num_pops, pop, forward_cont, backward_cont):
     args = [(i, num_pops, pop, backward_cont, forward_cont) for i in range(num_pops)]
 
     with Pool(cpu_count()) as pool:
-        r = list(
+        pools = list(
             tqdm(pool.imap(inside_loop_connectivity_matrices, args), total=len(args))
         )
         pool.close()
         pool.join()
 
     for i in range(num_pops):
[…]
