### 2020-02-01 524781e
cleaning up and recall performance
