+++ tools/plots.py
@@ -9,17 +9,17 @@ wave, inhibition, weights and noise.
 """
 
 import os
+import string
 
 import matplotlib.pyplot as plt
 import numpy as np
 from matplotlib.ticker import MaxNLocator
-import string
 
 FIG_DIR = "fig"
 os.makedirs(FIG_DIR, exist_ok=True)
 
 
-def currents(current_values, type_, fig_num, dt=1.0):
+def currents(current_values, type_, fig_num, t_step=1.0):
     """
     Plot current values.
 
@@ -29,7 +29,7 @@ def currents(current_values, type_, fig_num, dt=1.0):
     :param dt: float
     """
 
-    fig, ax = plt.subplots()
+    fig, axis = plt.subplots()
 
     titles = {
         "population": {
@@ -41,19 +41,19 @@ def currents(current_values, type_, fig_num, dt=1.0):
 
     n_iteration = current_values.shape[1]
 
-    x = np.arange(n_iteration, dtype=float) * dt
+    times = np.arange(n_iteration, dtype=float) * t_step
 
-    for i, y in enumerate(current_values):
-        ax.plot(x, y, linewidth=0.5, alpha=1)
+    for _, data in enumerate(current_values):
+        axis.plot(times, data, linewidth=0.5, alpha=1)
 
-    ax.set_xlabel("Time (cycles)")
-    ax.set_ylabel("Average current")
-    ax.set_title(titles[type_]["title"])
-    ax.text(
+    axis.set_xlabel("Time (cycles)")
+    axis.set_ylabel("Average current")
+    axis.set_title(titles[type_]["title"])
+    axis.text(
         -0.2,
         1.2,
         string.ascii_uppercase[fig_num],
-        transform=ax.transAxes,
+        transform=axis.transAxes,
         size=20,
         weight="bold",
     )
@@ -62,8 +62,10 @@ def currents(current_values, type_, fig_num, dt=1.0):
 
     plt.savefig(os.path.join(FIG_DIR, titles[type_]["filename"] + ".pdf"))
 
+    plt.close(fig)
 
-def firing_rates(f_rates, dt=1.0):
+
+def firing_rates(f_rates, t_step=1.0):
     """
     Plot firing rates.
 
@@ -71,24 +73,24 @@ def firing_rates(f_rates, dt=1.0):
     :param dt: float
     """
 
-    fig, ax = plt.subplots()
+    fig, axis = plt.subplots()
 
     n_iteration = f_rates.shape[1]
 
-    x = np.arange(n_iteration, dtype=float) * dt
-    ys = f_rates
+    times = np.arange(n_iteration, dtype=float) * t_step
+    n_data = f_rates
 
-    for i, y in enumerate(ys):
-        ax.plot(x, y, linewidth=0.5, alpha=1)
+    for _, data in enumerate(n_data):
+        axis.plot(times, data, linewidth=0.5, alpha=1)
 
-    ax.set_xlabel("Time (cycles)")
-    ax.set_ylabel("Average firing rate")
-    ax.set_title("Firing Rates")
-    ax.text(
+    axis.set_xlabel("Time (cycles)")
+    axis.set_ylabel("Average firing rate")
+    axis.set_title("Firing Rates")
+    axis.text(
         -0.2,
         1.2,
         string.ascii_uppercase[0],
-        transform=ax.transAxes,
+        transform=axis.transAxes,
         size=20,
         weight="bold",
     )
@@ -96,8 +98,10 @@ def firing_rates(f_rates, dt=1.0):
 
     plt.savefig(os.path.join(FIG_DIR, "firing_rates.pdf"))
 
+    plt.close(fig)
 
[…]
