### 2020-02-06 5f03ffb
parallelized net, not weights calculations
