+++ generate_seeds.py
@@ -8,6 +8,14 @@ import numpy as np
 
 np.random.seed(123)
 
-np.save(
-    os.path.join(".", "seeds"), np.random.randint(low=0, high=2 ** 32 - 1, size=10 ** 4)
-)
+
+def generate_seeds(force=False):
+    """ Generate and save 10**4 seeds, each to simulate one net """
+
+    if not os.path.exists("seeds.npy") or force:
+        print("Saving an array of seeds... ")
+        np.save(
+            os.path.join(".", "seeds"),
+            np.random.randint(low=0, high=2 ** 32 - 1, size=10 ** 4),
+        )
+        print("Done!")
