+++ main.py
@@ -24,7 +24,7 @@ from multiprocessing.pool import Pool
 import numpy as np
 from tqdm import tqdm
 
-import tools.plots as plot
+import generate_seeds
 import tools.recall_performance
 import tools.sine_wave
 
@@ -34,6 +34,7 @@ os.makedirs(FIG_DIR, exist_ok=True)
 
 np.seterr(all="raise")
 
+generate_seeds.generate_seeds()
 
 # === Parameters ===
 # Architecture
@@ -56,7 +57,7 @@ PHASE_SHIFT = 0.75
 CONT_FORTH = 1500
 CONT_BACK = 400
 # Time
-T_CYCLES = 8.0  # 14
+T_CYCLES = 1.0  # 14
 T_STEP = 0.001
 # Noise
 NOISE_VAR = 65
@@ -67,11 +68,8 @@ FIRST_MEMORY = 7
 RECALL_THRESHOLD = 15
 # Replication parameters
 PARAM_NOISE = 700
-# 10 we see saw peaks and bubbles and good transitions
-# print(param_noise)
-PARAM_INHIBITION = 1
-# print(param_inhibition)
 PARAM_CURRENT = 1  # 4.75
+
 
 # ==================
 
@@ -120,7 +118,7 @@ def compute_connectivity_matrices(num_pops, pop, forward_cont, backward_cont):
     return regular_connectivity, forward_connectivity, backward_connectivity
 
 
-def main():
+def main(arg):
     """ Licensing and main execution """
 
     print(
@@ -130,8 +128,10 @@ def main():
         "This is free software, and you are welcome to redistribute it\n"
         "under certain conditions; see the 'LICENSE' file for details.\n"
     )
+    # seed = 123, force = False, dump_pickle = True
+    seed, force, dump_pickle, CONT_FORTH = arg
 
-    np.random.seed(123)
+    np.random.seed(seed)
 
     # Compute memory patterns
     print("Computing memory patterns and neuron populations...")
@@ -186,29 +186,52 @@ def main():
             t_step=T_STEP,
         )
 
-    inhibition = -sine_wave * PARAM_INHIBITION
+    inhibition = -sine_wave
 
+    print(CONT_FORTH)
     # Compute weights
     print("Computing regular, forward and backward connectivity...")
 
-    bkp_file = "bkp/connectivities.p"
+    bkp_file = f"bkp/{seed}-connectivities.p"
 
     os.makedirs(os.path.dirname(bkp_file), exist_ok=True)
 
-    if not os.path.exists(bkp_file):
-        connectivities = compute_connectivity_matrices(
-            num_pops, pop, forward_cont, backward_cont
-        )
-        pickle.dump(connectivities, open(bkp_file, "wb"))
+    regular_connectivity = np.zeros((num_pops, num_pops))
+    forward_connectivity = np.zeros((num_pops, num_pops))
+    backward_connectivity = np.zeros((num_pops, num_pops))
 
-    else:
-        print("Loading connectivity from pickle file...")
-        connectivities = pickle.load(open(bkp_file, "rb"))
+    if not os.path.exists(bkp_file) or force:
+        # Compute weights
+        print("Computing regular, forward and backward connectivity...")
 
-    # regular_connectivity, forward_connectivity, backward_connectivity = \
-    #     compute_connectivity_matrices(num_pops, pop, forward_cont,
-    #                                   backward_cont)
-    regular_connectivity, forward_connectivity, backward_connectivity = connectivities
+        for i in tqdm(range(num_pops)):
+            for j in range(num_pops):
+                regular_connectivity[i, j] = np.sum(
+                    (pop[i, :] - SPARSITY)
+                    * (pop[j, :] - SPARSITY)
[…]
