### 2020-02-07 #10 @c-torre
Dear all,
Thank you for your patience so far. We had to scale up our computing and build a pipeline for the computer cluster due to very expensive computations (i.e. simulating many networks for many time steps, around 8h per network in the cluster). In that regard, we've been experiencing delays due to particulary long job queues in the system. We are working on this issue and expect to deliver our response in one week time.
