### 2020-02-09 92c9521
cummulative J
