### 2020-02-11 0905b19
corrected IRTs calculation
