+++ recall_analysis.py
@@ -10,13 +10,16 @@ import seaborn as sns
 PLOTS_DIR = "plt"
 RESULTS_DIR = "results"
 
+RESULTS_FILES = sorted(os.listdir(RESULTS_DIR))
+
 
 def get_cont_forth_name_values():
     # List each part of the file names separated by "-", get second element and remove "j"
     # This gets forward contiguity parameter value from file names
     return [
         int(file_name.split("-")[1][1:])
-        for file_name in sorted(os.listdir(RESULTS_DIR))
+        for file_name in RESULTS_FILES
+        if "j" in file_name
     ]
 
 
@@ -24,9 +27,12 @@ def get_recall_arrays():
 
     return [
         np.load(os.path.join(RESULTS_DIR, file_name))
-        for file_name in sorted(os.listdir(RESULTS_DIR))
-        if "j" and ".npy" in file_name
+        for file_name in RESULTS_FILES
+        if "j" in file_name
     ]
+
+
+#%%
 
 
 def get_cont_forth_data_frames():
@@ -99,26 +105,28 @@ ax.set(
 #%%
 
 
-def get_cumulative_recalls(recalls_data_frames)
-""" Build an array with average cumulative recall of the first recall of every recalled memory """
+def get_cumulative_recalls(recalls_data_frames):
+    """ Build an array with average cumulative recall of the first recall of recalled memories """
 
     cumulative_vectors = []
-    for recall_data_frame in recall_data_frames.values():
+    for recall_data_frame in recalls_data_frames.values():
         # Get first time index a memory is recalled, taking care of not recalled
         memory_first_recall_idx = pd.unique(recall_data_frame.idxmax())
+        # Pre-allocate vector
         cumulative_vector = np.zeros(recall_data_frame.shape[0])
         # Add 1 to all elements from the first recall until the end of array
         for memory_recall_idx in memory_first_recall_idx:
             cumulative_vector[memory_recall_idx:] += 1
-        cumulative_vectors.append(cummulative_vector)
-    
-    return np.average(np.vstack(cummulative_vectors), axis=0)
+        cumulative_vectors.append(cumulative_vector)
 
-cumulative_recalls = get_cumulative_recalls(recall_data_frames_su)
+    return np.average(np.vstack(cumulative_vectors), axis=0)
+
+
+cumulative_recalls = get_cumulative_recalls(recalls_data_frames_su)
 
 #%%
 
-ax = sns.lineplot(data=cummulative_recalls)
+ax = sns.lineplot(data=cumulative_recalls)
 ax.set(
     xlabel="Time (cycles)",
     ylabel="Average cumulative number of recalled memories",
@@ -127,40 +135,42 @@ ax.set(
 
 #%%
 
-recall_data_frames = recalls_data_frames_su[600]
+recalls_data_frames = recalls_data_frames_su[600]
 
 
-
-
-def get_memory_jumps(recalls):
+def get_memory_jumps_idx(recalls):
     """ When the recalled memory changes """
 
-    # Get a column array of the memory indexes
-    memory_identifier = np.rot90(np.arange(recalls.shape[0]))
-    # Multiply binary array by previous to identify the memories, flatten
-    recalled_memories_time_list = np.sum((recalls * memory_identifier), axis=1)
-    # Substract one element from the previous to get the size of memory jumps
-    memory_jump_sizes = np.diff(recalled_memories_time_list)
+    recalls = recalls_data_frames
+    # Get when there is a change in the recalled memory, a jump
+    jumps = recalls.diff()
+    # Eliminate all -1 from diff() to get ones after sum, where jumps happened
+    jumps = jumps[jumps == 1]
[…]
