+++ model.py
@@ -1,3 +1,4 @@
+#%%
 """
 <[Re] Recanatesi (2015). Neural Network Model of Memory Retrieval>
 Copyright (C) <2019>  <de la Torre-Ortiz C, Nioche A>
@@ -140,20 +141,20 @@ def main(arg):
     )
 
     # Compute populations
-    pop, neurons_per_pop = np.unique(
+    pop_per_memory, neurons_per_pop = np.unique(
         [tuple(i) for i in memory_patterns], axis=0, return_counts=True
     )
 
-    num_pops = len(pop)
+    num_pops = len(pop_per_memory)
 
-    neurons_encoding_mem = [
-        (pop[:, mu] == 1).nonzero()[0] for mu in range(NUM_MEMORIES)
-    ]
+    pop_num_encoding_mem = [
+        (pop_per_memory[:, memory] == 1).nonzero()[0] for memory in range(NUM_MEMORIES)
+    ]  # Not used in this version of the model
 
     # === Other pre-computations ===
     num_iter = int(T_CYCLES / T_STEP)
     # relative_excitation = excitation / NUM_NEURONS
-    time_param = T_STEP / T_DECAY
+    time_tot = T_STEP / T_DECAY
 
     # Inhibition
     sine_wave = np.zeros(num_iter)
@@ -188,7 +189,6 @@ def main(arg):
 
     inhibition = -sine_wave
 
-    print(CONT_FORTH)
     # Compute weights
     print("Computing regular, forward and backward connectivity...")
 
@@ -207,15 +207,15 @@ def main(arg):
         for i in tqdm(range(num_pops)):
             for j in range(num_pops):
                 regular_connectivity[i, j] = np.sum(
-                    (pop[i, :] - SPARSITY) * (pop[j, :] - SPARSITY)
+                    (pop_per_memory[i, :] - SPARSITY) * (pop_per_memory[j, :] - SPARSITY)
                 )
 
                 forward_connectivity[i, j] = np.sum(
-                    pop[i, forward_cont] * pop[j, forward_cont + 1]
+                    pop_per_memory[i, forward_cont] * pop_per_memory[j, forward_cont + 1]
                 )
 
                 backward_connectivity[i, j] = np.sum(
-                    pop[i, backward_cont] * pop[j, backward_cont - 1]
+                    pop_per_memory[i, backward_cont] * pop_per_memory[j, backward_cont - 1]
                 )
 
         # if dump_pickle:
@@ -242,21 +242,21 @@ def main(arg):
     # Compute noise
     print("Calculating uncorrelated Gaussian noise...")
 
-    for pop in range(num_pops):
-        noise[pop] = (
+    for pop_per_memory in range(num_pops):
+        noise[pop_per_memory] = (
             np.random.normal(
-                loc=0, scale=(NOISE_VAR * neurons_per_pop[pop]) ** 0.5, size=num_iter
+                loc=0, scale=(NOISE_VAR * neurons_per_pop[pop_per_memory]) ** 0.5, size=num_iter
             )
-            / neurons_per_pop[pop]
+            / neurons_per_pop[pop_per_memory]
             * PARAM_NOISE
         )  # MOD ADDED BACK
 
     # Initialize firing rates
-    firing_rates[neurons_encoding_mem[FIRST_MEMORY]] = INIT_RATE
+    firing_rates[pop_num_encoding_mem[FIRST_MEMORY]] = INIT_RATE
 
     # Initialize current
-    c_ini = INIT_RATE ** (1 / GAIN_EXP) - GAIN_THRESHOLD
-    current[neurons_encoding_mem[FIRST_MEMORY]] = c_ini
[…]
