### 2020-02-14 f841f9c
cleaned data load, new results
