+++ recall_performance/data_build.py
@@ -1,200 +1,0 @@
-#%%
-import os
-
-import numpy as np
-import pandas as pd
-from tqdm import tqdm
-
-from recall_performance import data_loader
-
-PLOTS_DIR = "plt"
-RESULTS_DIR = "results"
-
-
-# Data load and clean
-SEEDS_RESULTS_DIR = os.path.join(RESULTS_DIR, "seeds")
-
-
-recalls_data_frames = data_loader.get_cleaned_frames(SEEDS_RESULTS_DIR)
-
-
-# Avoid 0/0 error with hacky sum later
-EPSILON = 0.000000001
-# recalls_data_frames = list(npy_loader_cont_forth.get_cont_forth_data_frames().values())
-
-# recalls_data_frames = [
-#     drop_bad_recalls(recall_data_frame) for recall_data_frame in recalls_data_frames
-# ]
-
-
-# example_frame = recalls_data_frames[5]
-
-#%%
-
-
-def make_master_data_frame(recalls_data_frame):
-    """ Create the data frame that contains most information for plots """
-
-    return pd.DataFrame(index=range(len(recalls_data_frame.index)))
-
-
-# recall_analysis_data_frame = make_master_data_frame()
-
-
-def _get_memory_jumps_idx(recalls):
-    """ When the recalled memory changes """
-
-    # Get when there is a change in the recalled memory, a jump
-    jumps = recalls.diff()
-    # Eliminate all -1 from diff() to get ones after sum, where jumps happened
-    jumps = jumps[jumps == 1]
-    # Sum along columns to obtain the vector of jumps, 1 when jumped, 0 when not
-    jumps_vector = jumps.sum(axis=1)
-    # Add its index information to the memory jump
-    jumps_idx = jumps_vector * np.arange(jumps_vector.size)
-    # Remove zeros to just return indices
-    return jumps_idx.nonzero()[0]
-
-
-# example_jumps = get_memory_jumps_idx(example_frame)
-
-
-def _get_inter_retrieval_times(jumps_idx):
-    """
-    IRT is the number of iterations that pass before a new memory is recalled.
-    e.g. p1, p1, p1, p5; IRT = 3
-    """
-
-    # There needs to be an imaginary jump just before first memory recall to compute IRTs
-    first_jump_idx = [-1]
-    # Stack an imaginary first jump to compute diffs correctly
-    jumps_idx_probed = np.hstack((first_jump_idx, jumps_idx))
-    # Diffs of jump indexes gives iterations until jump; these are the IRT
-    return np.diff(jumps_idx_probed)
-
-
-# example_irts = get_inter_retrieval_times(example_jumps)
-
-
-def _expand_irts(irts, recalls_data_frame):
-    """
-    Transform a group of IRTs into its time series.
-    e.g. (3, 2, 1) to (0, 0, 3, 0, 2, 1)
-    """
-
-    if irts.size != 0:
-        return np.hstack([(np.hstack((np.zeros(irt - 1), irt))) for irt in irts])
-    else:
-        return np.zeros(recalls_data_frame.shape[0])
-
-
-def get_irts_time_series(recalls):
-    """ Get time indexed inter-retrieval times """
-
-    jumps_idx = _get_memory_jumps_idx(recalls)
-    inter_retrieval_times = _get_inter_retrieval_times(jumps_idx)
-    return pd.Series(_expand_irts(inter_retrieval_times, recalls), name="irt")
-
-
-#%%
-
-
-def get_vectorized_probed_recalls(recalls_data_frame):
-    # Duplicate?
-
-    probe = np.cumsum(np.ones_like(recalls_data_frame), axis=1) - 1
[…]
