### 2020-02-14 0a2653b
cleaning, loops for all data
