+++ recall_analysis/data_processing/memory_intersections.py
@@ -3,92 +3,39 @@ from itertools import combinations
 
 import numpy as np
 import pandas as pd
+from tqdm import tqdm
 
 import paths
-import os
+import recall_analysis.data_processing.data_loader
 
-ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
-#
-RESULTS_DIR = os.path.join(ROOT_DIR, "results")
-##
-CONT_RESULTS_DIR = os.path.join(RESULTS_DIR, "forward_contiguity")
-##
-NOISE_RESULTS_DIR = os.path.join(RESULTS_DIR, "noise")
-##
-SEED_RESULTS_DIR = os.path.join(RESULTS_DIR, "seed")
-###
-POPULATION_SIZES_DIR = os.path.join(SEED_RESULTS_DIR, "population_sizes")
-###
-POPULATIONS_MEMORIES_DIR = os.path.join(SEED_RESULTS_DIR, "populations_memories")
-###
-RECALLS_SEED_DIR = os.path.join(SEED_RESULTS_DIR, "recalls")
-
-#%%
-
-np.seterr(all="raise")
-
-
-# === Parameters ===
-# Architecture
-NUM_NEURONS = 10 ** 5
-NUM_MEMORIES = 16
-# Hebbian rule
-SPARSITY = 0.1
-# ==================
-
-
-def main(arg):
-    seed = arg
-
-    np.random.seed(seed)
-
-    # Compute memory patterns
-    memory_patterns = np.random.choice(
-        [0, 1], p=[1 - SPARSITY, SPARSITY], size=(NUM_NEURONS, NUM_MEMORIES)
-    )
-
-    # Compute populations
-    pop, neurons_per_pop = np.unique(
-        [tuple(i) for i in memory_patterns], axis=0, return_counts=True
-    )
-
-    num_pops = len(pop)
-
-    return pop, neurons_per_pop
-    # np.save(os.path.join("results", f"s{seed}-neurons-encoding"), neurons_encoding_mem)
-
-
-from tqdm import tqdm
-def get_memory_data():
-    seeds = np.load("seeds.npy")
-    seeds = seeds[:105]
-    for seed in tqdm(seeds):
-        (pops), (neurons_per_pop) = main(seed)
-        np.save(
-            os.path.join(POPULATIONS_MEMORIES_DIR, f"s{seed}-populations-memories"),
-            pops,
-        )
-        np.save(
-            os.path.join(POPULATION_SIZES_DIR, f"s{seed}-population-sizes"),
-            neurons_per_pop,
-        )
-    return (
-        pd.DataFrame(pops),
-        pd.Series(neurons_per_pop),
-    )
+POPULATIONS_MEMORIES_DIR = paths.POPULATIONS_MEMORIES_DIR
+POPULATION_SIZES_DIR = paths.POPULATION_SIZES_DIR
 
 
 #%%
+
+populations_memories = recall_analysis.data_processing.data_loader.get_arrays_from_files(
+    POPULATIONS_MEMORIES_DIR
+)
+
+populations_memories = recall_analysis.data_processing.data_loader.arrays_to_data_frames(
+    populations_memories
+)
+
+population_sizes = recall_analysis.data_processing.data_loader.get_arrays_from_files(
+    POPULATION_SIZES_DIR
+)
 
 # Data load
-((pops), (neurons_per_pop)) = get_memory_data()
+# ((pops), (neurons_per_pop)) = get_memory_data()
 
 #%%
 
-def make_all_possible_intersections():
+
+def make_all_possible_intersections(populations_memories):
 
     # Get memory indices
-    num_memories = pops.shape[1]
+    num_memories = populations_memories.shape[1]
     memories_idx = np.arange(num_memories)
[…]
