### 2020-02-17 264d3ef
refactoring, architecture, fixing figs
