+++ recall_analysis/data_processing/pre_processing/intersections.py
@@ -1,4 +1,9 @@
+"""
+
+"""
+
 #%%
+
 from itertools import combinations
 
 import numpy as np
@@ -10,16 +15,6 @@ from recall_analysis.data_processing.pre_processing import loader_utils
 
 POPULATIONS_MEMORIES_DIR = paths.POPULATIONS_MEMORIES_DIR
 POPULATION_SIZES_DIR = paths.POPULATION_SIZES_DIR
-
-
-#%%
-
-populations_memories = loader_utils.get_arrays_from_files(POPULATIONS_MEMORIES_DIR)
-
-populations_memories = loader_utils.arrays_to_data_frames(populations_memories)
-
-population_sizes = loader_utils.get_arrays_from_files(POPULATION_SIZES_DIR)
-
 # Data load
 # ((pops), (neurons_per_pop)) = get_memory_data()
 
@@ -86,11 +81,15 @@ def get_intersection_sizes(intersections_data_frame, population_sizes):
 
 def make_all():
 
+    populations_memories = loader_utils.get_arrays_from_files(POPULATIONS_MEMORIES_DIR)
+    populations_memories = loader_utils.arrays_to_data_frames(populations_memories)
+    population_sizes = loader_utils.get_arrays_from_files(POPULATION_SIZES_DIR)
+
     assert len(populations_memories) == len(population_sizes)
 
     intersection_sizes_all = []
 
-    print("Building all data frames for recalls analysis...")
+    print("Building all series for memory intersection size...")
     for idx in tqdm(range(len(population_sizes))):
         intersections_data_frame = make_all_possible_intersections(
             populations_memories[idx]
@@ -103,5 +102,5 @@ def make_all():
     return intersection_sizes_all
 
 
-if __name__ == "__main__":
-    intersections_sizes_all = make_all()
+# if __name__ == "__main__":
+#     intersections_sizes_all = make_all()
