### 2020-02-17 5dc8282
fix rank transitions fig
