### 2020-02-17 adde3bb
plot fixes
