+++ recall_analysis/data_processing/pre_processing/recalls.py
@@ -66,8 +66,7 @@ def _expand_irts(irts, recalls_data_frame):
 
     if irts.size != 0:
         return np.hstack([(np.hstack((np.zeros(irt - 1), irt))) for irt in irts])
-    else:
-        return np.zeros(recalls_data_frame.shape[0])
+    return np.zeros(recalls_data_frame.shape[0])
 
 
 def get_irts_time_series(recalls):
@@ -87,18 +86,20 @@ def get_vectorized_probed_recalls(recalls_data_frame):
     probe = np.cumsum(np.ones_like(recalls_data_frame), axis=1) - 1
     probed_recalls = recalls_data_frame * (probe)
     vectorized_probed_recalls = probed_recalls.sum(axis=1)
+    print(vectorized_probed_recalls)
     return vectorized_probed_recalls
 
 
 #%%
 
 
-def get_jumps(vectorized_recalls):
+def get_transitions(vectorized_recalls):
     # Duplicate?
 
-    jumps_absolute = vectorized_recalls.diff().dropna()
-    jumps_bool = (jumps_absolute != 0).astype(int)
-    return pd.Series(jumps_bool, name="transition", dtype=bool)
+    jumps_absolute = vectorized_recalls.diff()  # .fillna(0)
+    jumps_bool = jumps_absolute != 0
+    jumps_bool.loc[0] = False
+    return pd.Series(jumps_bool, name="transition")
 
 
 #%% Hunting the first appereance
@@ -117,21 +118,66 @@ def get_first_unique_recalls(vectorized_probed_recalls, recalls_analysis_data_fr
     )
 
     # Set an array with ones in the indices of unique recall
-    first_appereances = pd.Series(np.zeros(recalls_analysis_data_frame.shape[0]))
-    first_appereances.iloc[unique_memory_recalls_idx] = 1
+    first_appereances = pd.Series(
+        np.zeros(recalls_analysis_data_frame.shape[0]), dtype=bool
+    )
+    first_appereances.iloc[unique_memory_recalls_idx] = True
     return pd.Series(first_appereances, name="unique_recalls")
 
 
 def get_counts_unique_recalls(first_appereances):
     # Get counts from first appereances in previous array
     counter = 1
-    counts_first_appereances = np.zeros_like(first_appereances)  # X
+    counts_first_appereances = np.zeros_like(first_appereances, dtype=int)  # X
     for idx, element in enumerate(first_appereances):
-        if element == 1:
+        if element:
             counts_first_appereances[idx] = counter
             counter += 1
 
     return pd.Series(counts_first_appereances, name="unique_recall_count")
+
+
+def get_memory_sizes(data_frame_idx):
+    # Get data
+    populations_memories_dir = paths.POPULATIONS_MEMORIES_DIR
+    population_sizes_dir = paths.POPULATION_SIZES_DIR
+
+    populations_memories = loader_utils.get_arrays_from_files(
+        populations_memories_dir, calling_intersections=True
+    )
+    populations_memories = loader_utils.arrays_to_data_frames(populations_memories)
+    population_sizes = loader_utils.get_arrays_from_files(
+        population_sizes_dir, calling_intersections=True
+    )
+
+    # Get memory sizes by multiplying each population by its size and adding all beloging to memory
+    populations_with_sizes = populations_memories[data_frame_idx] * np.rot90(
+        np.expand_dims(population_sizes[data_frame_idx], axis=0), k=3
+    )
+
+    memory_sizes = populations_with_sizes.sum()
+    memory_sizes.name = "memory_size"
+    return memory_sizes
+
+
+def merge_recalled_memory_with_size(data_frame_idx, recalled_memories):
+
+    # recalls_frame = recalls_frame["memory_recalled"]
[…]
