+++ results/data_info.md
@@ -1,0 +1,2 @@
+
+names delicate as separator and number of elements used in computations

[…]
