+++ recall_analysis/data_processing/pre_processing/recalls.py
@@ -11,6 +11,16 @@ from tqdm import tqdm
 import paths
 from recall_analysis.data_processing.pre_processing import loader_utils
 
+populations_memories_dir = paths.POPULATIONS_MEMORIES_DIR
+population_sizes_dir = paths.POPULATION_SIZES_DIR
+
+populations_memories = loader_utils.get_arrays_from_files(
+    populations_memories_dir, calling_intersections=True
+)
+populations_memories = loader_utils.arrays_to_data_frames(populations_memories)
+population_sizes = loader_utils.get_arrays_from_files(
+    population_sizes_dir, calling_intersections=True
+)
 #%%
 
 
@@ -86,17 +96,33 @@ def get_vectorized_probed_recalls(recalls_data_frame):
     probe = np.cumsum(np.ones_like(recalls_data_frame), axis=1) - 1
     probed_recalls = recalls_data_frame * (probe)
     vectorized_probed_recalls = probed_recalls.sum(axis=1)
-    vectorized_probed_recalls = vectorized_probed_recalls.astype(int)
-    sanitize = vectorized_probed_recalls.values
-    sanitize = sanitize.astype(int)
-    sanitize = list(sanitize)
-    sanitize = [int(elem) for elem in sanitize]
-    # print(vectorized_probed_recalls)
-    print(vectorized_probed_recalls.values)
-    returning = pd.Series(vectorized_probed_recalls, name="recalled_memory", dtype=int)
-    # assert returning.values[5] is int
-    return returning
+    sanitize = (
+        vectorized_probed_recalls.values
+    )  # Be warned, Gods of chaos awakened if float
+    sanitize = pd.Series(sanitize, name="recalled_memory", dtype=int)
+    return sanitize
 
+
+# def get_vectorized_probed_recalls(recalls_data_frame):
+#     # Duplicate?
+
+#     probe = np.cumsum(np.ones_like(recalls_data_frame), axis=1) - 1
+#     probed_recalls = recalls_data_frame * (probe)
+#     vectorized_probed_recalls = probed_recalls.sum(axis=1)
+#     vectorized_probed_recalls = vectorized_probed_recalls.astype(int)
+#     sanitize = vectorized_probed_recalls.values
+#     sanitize = sanitize.astype(int)
+#     sanitize = list(sanitize)
+#     sanitize = [str(elem) for elem in sanitize]
+#     sanitize = [int(elem) for elem in sanitize]
+#     # print(vectorized_probed_recalls)
+#     # print(vectorized_probed_recalls.values)
+#     returning = pd.Series(sanitize, name="recalled_memory", dtype=int)
+#     # assert returning.values[5] is int
+#     # print(returning.values)
+#     # returning = pd.to_numeric(returning, errors="raise", downcast="integer")
+#     print(sanitize)
+#     return sanitize
 
 #%%
 
@@ -147,16 +173,6 @@ def get_counts_unique_recalls(first_appereances):
 
 def get_memory_sizes(data_frame_idx):
     # Get data
-    populations_memories_dir = paths.POPULATIONS_MEMORIES_DIR
-    population_sizes_dir = paths.POPULATION_SIZES_DIR
-
-    populations_memories = loader_utils.get_arrays_from_files(
-        populations_memories_dir, calling_intersections=True
-    )
-    populations_memories = loader_utils.arrays_to_data_frames(populations_memories)
-    population_sizes = loader_utils.get_arrays_from_files(
-        population_sizes_dir, calling_intersections=True
-    )
 
     # Get memory sizes by multiplying each population by its size and adding all beloging to memory
     populations_with_sizes = populations_memories[data_frame_idx] * np.rot90(
@@ -181,9 +197,9 @@ def merge_recalled_memory_with_size(data_frame_idx, recalled_memories):
         left_on=[recalled_memories],
[…]
