+++ recall_analysis/plots/irts_transition_number.py
@@ -66,10 +66,7 @@ indexed = [
 
 #%%
 indexed = [
-    pd.Series(
-        series.values[series.values.nonzero()[0]],
-        name=series.name
-    )
+    pd.Series(series.values[series.values.nonzero()[0]], name=series.name)
     for series in indexed
 ]
 #%%
@@ -91,128 +88,195 @@ spam = pd.DataFrame(indexed)
 #%%
 
 eggs = spam.groupby(by=spam.index).sum().reset_index(drop=True)
+
 #%%
+
 eggs /= len(recalls_analysis_data_frames_all)
 eggs.index += 1
 eggs.index.name = "average_irt_unique_transition"
+
 #%%
+
 eggerino = eggs.dropna(axis=1, how="all")  # change me to replace by 0s
 eggerino = eggerino.loc[:, (eggerino != 0).any(axis=0)]  # Drop column only 0
+
 #%%
 # pd.concat([eggs], keys=['foo'], names=['Firstlevel'])
 
+# #%%
+
+# example_frame = recalls_analysis_data_frames_all[5]
+
+# branches_plot_data = pd.DataFrame(
+#     index=range(len(example_frame.index)),
+#     columns=["new_recall_jumps", "jumps_cum_sum", "irts_cum_sum"],
+# )
+
+
+# def get_memory_jumps_idx(recalls):
+#     """ When the recalled memory changes """
+
+#     # Get when there is a change in the recalled memory, a jump
+#     jumps = recalls.diff()
+#     # Eliminate all -1 from diff() to get ones after sum, where jumps happened
+#     jumps = jumps[jumps == 1]
+#     # Sum along columns to obtain the vector of jumps, 1 when jumped, 0 when not
+#     jumps_vector = jumps.sum(axis=1)
+#     # Add its index information to the memory jump
+#     jumps_idx = jumps_vector * np.arange(jumps_vector.size)
+#     # Remove zeros to just return indices
+#     return jumps_idx.nonzero()[0]
+
+
+# example_jumps = get_memory_jumps_idx(example_frame)
+
+# #%%
+
+
+# def get_inter_retrieval_times(jumps_idx):
+#     """
+#     IRT is the number of iterations that pass before a new memory is recalled.
+#     e.g. p1, p1, p1, p5; IRT = 3
+#     """
+
+#     # There needs to be an imaginary jump just before first memory recall to compute IRTs
+#     first_jump_idx = [-1]
+#     # Stack an imaginary first jump to compute diffs correctly
+#     jumps_idx_probed = np.hstack((first_jump_idx, jumps_idx))
+#     # Diffs of jump indexes gives iterations until jump; these are the IRT
+#     return np.diff(jumps_idx_probed)
+
+
+# example_irts = get_inter_retrieval_times(example_jumps)
+# #%% Building the array for the multi gradient line plot
+
+
+# def expand_irts(irts):
+#     """
+#     Transform a group of IRTs into its time series.
+#     e.g. (3, 2, 1) to (0, 0, 3, 0, 2, 1)
+#     """
+
+#     return np.hstack([(np.hstack((np.zeros(irt - 1), irt))) for irt in irts])
+
+
+# irts_time_series = expand_irts(example_irts)  # Y num
+# irts_cum_sum = pd.Series(irts_time_series.cumsum())
+# branches_plot_data.loc[:, "irts_cum_sum"] = irts_cum_sum
+
+# # #%%
+
+# probe = np.cumsum(np.ones_like(example_frame), axis=1) - 1
+# probed_recalls = example_frame * (probe)
+# vectorized_probe_recalls = probed_recalls.sum(axis=1)
+# jumps_absolute = vectorized_probe_recalls.diff().dropna()
[…]
