### 2020-03-02 c918838
just before making the loop
