+++ model/trials.py
@@ -1,18 +1,25 @@
 #%%
-import pandas as pd
-import numpy as np
-from tqdm import tqdm
-import seaborn as sns
 import matplotlib.pyplot as plt
+import matplotlib.style as mpstyle
+import numpy as np
+import pandas as pd
+import seaborn as sns
+from tqdm import tqdm
 
 np.random.seed(123)
 
 tqdm.pandas()
 
-NUM_NEURONS = 100000
+mpstyle.use("fast")
+
+
+NUM_NEURONS = 10 ** 5
 NUM_MEMORIES = 16
-T_STEP = 0.01
-T_TOT = 10
+# Activation
+T_DECAY = 0.01
+# Time
+T_STEP = 0.001
+T_TOT = 5
 T_SIMULATED = int(T_TOT / T_STEP)
 # Hebbian rule
 EXCITATION = 13_000
@@ -21,10 +28,183 @@ SPARSITY = 0.1
 GAIN_THRESHOLD = 0
 GAIN_EXP = 2 / 5
 # Inhibition
-SIN_MIN = 0.7
-SIN_MAX = 1.0
+SIN_MIN = 0.7 * EXCITATION
+SIN_MAX = 1.06 * EXCITATION
 # Noise
 NOISE_STD = 65
+# Forward and backward contiguity
+CONT_FORTH = 400 / EXCITATION
+CONT_BACK = 1500 / EXCITATION
+# R
+NUM_POINTS = T_TOT * 100
+
+#%%
+# >>>>>>>>>>>>>>>>>>>>>>> Reca
+
+population_sizes = np.array(NUM_NEURONS)
+log_int = np.array(0)
+sparsity_vect = np.ones(NUM_MEMORIES) * SPARSITY
+
+#%%
+
+
+def build_logical(prev_logical):
+    """
+    Stacks the previous logical array to itself, then adds an extra row at the bottom,
+    with half zeros and half ones
+    R: chaging this line practically does meanfield 
+    """
+
+    composite = np.hstack((prev_logical, prev_logical))
+    try:
+        half_size = prev_logical.shape[1]
+    except:
+        half_size = 1
+    bottom = np.hstack((np.zeros(half_size), np.ones(half_size)))
+    composite = np.vstack((composite, bottom))
+    # print(composite)
+    return composite
+
+
+for i in range(NUM_MEMORIES):
+    new_sizes = (
+        population_sizes * SPARSITY
+    )  # R: calculates the sizes of the new intersctions
+    population_sizes = np.hstack(
+        ((population_sizes - new_sizes), new_sizes)
+    )  # R: updates the sizes
+
+    log_int = build_logical(log_int)
+
+#%%
+
+pop_siz = population_sizes.T  # R: final vector of sizes
+
+num_pops = population_sizes.shape[0]  # R: size of the actual simulated network
+
+#%%
+
+connectivity_back = (log_int.T)[
+    :, :16
+]  # R: final logical identity matrix; this is correct, not reg
+connectivity_reg = np.hstack(
+    (connectivity_back[:, 1:], np.zeros(connectivity_reg.shape[0])[:, None])
+)
+connectivity_forth = np.hstack(
+    (connectivity_reg[:, 1:], np.zeros(connectivity_reg.shape[0])[:, None])
+)
+
+#%%
+
+it_rec = np.linspace(1, T_SIMULATED, NUM_POINTS).astype(int)
+time = np.arange(start=1, stop=T_TOT + 1 + T_STEP, step=T_STEP)
+#%%
+
+rates = np.zeros((num_pops, len(it_rec)))
+
+
+#%%
+
+t_rates = (
+    time[it_rec - 1] - 1
[…]
