### 2020-03-09 #10 @oliviaguest
@rougier this seems to be a reason to rewrite and re-review this work? What's our policy on this type of situation?
