### 2020-04-07 f48aa78
added proofs
