+++ model/proofs.py
@@ -1,0 +1,255 @@
+#%%
+import os
+import pickle
+
+import numpy as np
+import pandas as pd
+from tqdm import tqdm
+
+np.random.seed(123)
+
+
+NUM_NEURONS = 1  # 0 ** 5
+NUM_MEMORIES = 16
+# Activation
+T_DECAY = 0.01
+# Time
+T_STEP = 0.001
+T_TOT = 5
+T_SIMULATED = int(T_TOT / T_STEP)
+# Hebbian rule
+EXCITATION = 13_000
+SPARSITY = 0.1
+# Gain
+GAIN_THRESHOLD = 0
+GAIN_EXP = 2 / 5
+# Inhibition
+SIN_MIN = 0.7 * EXCITATION
+SIN_MAX = 1.06 * EXCITATION
+# Noise
+NOISE_STD = 65
+# Forward and backward contiguity
+CONT_FORTH = 1500 / EXCITATION
+CONT_BACK = 400 / EXCITATION
+# R
+NUM_POINTS = T_TOT * 100
+
+ITERS = 3  # CHANGE ME
+
+#%% a
+
+
+def get_log_int_sizes():
+    def build_logical(prev_logical):
+        """
+        Stacks the previous logical array to itself, then adds an extra row at the bottom,
+        with half zeros and half ones
+        R: chaging this line practically does meanfield 
+        """
+
+        composite = np.hstack((prev_logical, prev_logical))
+        try:
+            half_size = prev_logical.shape[1]
+        except:
+            half_size = 1
+        bottom = np.hstack((np.zeros(half_size), np.ones(half_size)))
+        composite = np.vstack((composite, bottom))
+
+        return composite
+
+    # (1,)
+    population_sizes = np.array(NUM_NEURONS)
+    # (1,)
+    log_int = np.array(0)
+
+    for i in range(NUM_MEMORIES):
+        # print("pop sizes:", population_sizes)
+        new_sizes = (
+            population_sizes * SPARSITY
+        )  # R: calculates the sizes of the new intersctions
+        # print("new sizes:", new_sizes)
+        population_sizes = np.hstack(
+            ((population_sizes - new_sizes), new_sizes)
+        )  # R: updates the sizes
+        # print("pop sizes:", population_sizes)
+
+        log_int = build_logical(log_int)
+
+        # print("=" * 10)
+        if i == ITERS:
+            break
+
+    # (65536,), (17, 65536)
+    return population_sizes, log_int
+
+
+# pop_siz = population_sizes.T  # R: final vector of sizes
+
+
+#%%
+
+
+def get_connectivities(log_int):
+    # ]  # R: final logical identity matrix
+    # (65536, 16)
+    connectivity_reg = (log_int.T)[:, 1:]
+    # (65536, 16)
+    connectivity_back = np.hstack(
+        (
+            np.zeros(connectivity_reg.shape[0])[:, None],
+            connectivity_reg[:, : NUM_MEMORIES - 1],
+        )
+    )
+    # (65536, 16)
+    connectivity_forth = np.hstack(
+        (connectivity_reg[:, 1:], np.zeros(connectivity_reg.shape[0])[:, None])
+    )
+    # 3 x (65536, 16)
+    return connectivity_reg, connectivity_back, connectivity_forth
+
+
+def prepare_times():
+    # (500,)
+    it_rec = np.linspace(1, T_SIMULATED, NUM_POINTS).astype(int)
+    # (5001,)
+    time = np.arange(start=1, stop=T_TOT + 1 + T_STEP, step=T_STEP)
+    # (500,)
[…]

[…]
