### 2020-05-04 #10 @c-torre
Dear everyone,
I hope you are safe in these difficult times.

In this time, we addressed the reviewer comments, including the change in one parameter value.
The model shown to be particularly sensitive to this and the dynamics were completely distorted.

We then reimplemented the model making new adjustments to equations and parameters, as we were able to obtain figure 2 again.
We launched series of simulations to reproduce the last figures of the original article.
Unfortunately, results were not as expected.

We contacted the original author, who provided apparently a similar implementation to what they used for their article.
They also provided further explanations on theoretical concepts driving their implementaiton.
After reimplementing their code in Python, we obtained yet another working version of the network that could replicate figure 2.
We launched the extended simulations to produce the last figures, but again its behavior was far from intended.
We were not able to get in touch with the original author again.

On the other hand, we are now able to provide a lot more details on the theory and implementation.
In addition, we have two working versions of the code driven by similar but not equal theory:
we have a low sparsity version that works with simulation numbers as described in the article;
we also have a high sparsity version assumes an infinite number of neurons, which is the closest to the original implementation.

We are wondering how we should proceed in this case.
