### 2020-05-04 #10 @oliviaguest
> We contacted the original author, who provided apparently a similar implementation to what they used for their article.

> We were not able to get in touch with the original author again.

@c-torre Can you please provide a bit more info regarding what has changed and why? I don't blame you in any way, I just would like to know more and your opinion.
