### 2020-05-04 #10 @c-torre
The complete timeline occured as follows:
I contacted the corresponding author of the original article, but I did not get a reply back.
Then I contacted the first author. He was very willing to help, provided some initial discussion and finally the partial code that allowed to reproduce figure 2.

I launched the longer simulations to get results for the last figures, but again could not reproduce results. At this point I got back to contact the first author of the original article. I could not get a reply even after reminding.

I'm wondering if the current situation may have an effect in any way, but unfortunately it is all guessing just by now. It does not seem to be any other changes that would explain why we could not get back to them again.
