### 2020-05-04 #10 @oliviaguest
@rougier @khinsen @benoit-girard what do we do/think in this case? Is there a value in asking for input from (any of) the original authors?
