### 2020-05-05 #10 @rougier
**Disclaimer**: I've published articles with the second author of this submission (@AurelienNioche)

From my understanding, authors managed to replicate original results using the low sparsity version with (heavy ?) modification of the original model. I think it is thus a successful replication even though the model seems very fragile and this must be stated somewhere in the paper. Since the corresponding authors is now un-responsive, I'm not sure it's worth contacting the other authors. Or we set a hard deadline for getting an answer (2 weeks?) and if they don't get the answer, we can publish the paper (provided reviewers and editor agree).
