### 2020-05-05 #10 @oliviaguest
I agree on all points, but I think it's worth contacting all the authors of the original paper since if I was one of them I would want to know?
