### 2020-05-05 #10 @khinsen
I cannot comment on the changes to the models, that's beyond my domain competence (which is close to zero). As @oliviaguest says, as an author of the original work, I'd like to be informed, but I'd say "informing" is sufficient - no need to wait for a reply that may or may not come.

There is also the question of whether this is a successful replication or not. I'd say it's neither, but we have known for a while that our success/failure classification does not cover all situations. My understanding of this thread is that results similar to the original publication could only be obtained by modifying the underlying model. This suggests one of (1) the original implementation has a problem, (2) the re-implementation has a problem, (3) both implementations have a problem, or (4) the model is too sensitive to implementation details to be reproducible in practice.

The ideal way to proceed (after accepting this replication of course) would be to encourage the authors of original and replication to work together on figuring out what happened.
