### 2020-05-05 #10 @oliviaguest
I agree on no delay — sorry that I left that ambiguous — by waiting on anybody. Merely that I think we should inform them and maybe wait a bit (because they might give us useful info), but certainly not an endless amount of time.

> The ideal way to proceed (after accepting this replication of course) would be to encourage the authors of original and replication to work together on figuring out what happened.

I am happy with this idea. I wonder what @c-torre would prefer?
