### 2020-05-14 ba77a7d
cleaned katkov's script
