+++ model/mkatkov.py
@@ -1,6 +1,5 @@
 #%%
 import numpy as np
-import scipy
 import scipy.sparse
 import matplotlib.pyplot as plt
 from tqdm import tqdm
@@ -11,93 +10,116 @@ from IPython.display import display, clear_output
 
 class Model:
     def __init__(self):
-        self.th0 = 0
-        self.gamma = 0.5
-        self.N = 100000
-        self.P = 16
-        self.f = 0.01
-        self.kappa = 11
-        self.xi0 = 1
-        self.dt = 1e-2
+        self.GAIN_THRESHOLD = 0
+        self.GAIN_EXPONENT = 0.5
+        self.NUM_NEURONS = 10 ** 5
+        self.NUM_MEMORIES = 16
+        self.SPARSITY = 0.01
+        self.EXCITATION = 11
+        self.NOISE_VAR = 1
+        self.T_STEP = 1e-2
         self.T = 10
-        self.phi = 0
-        self.tau = 1e-1
+        self.OSCILLATION = 0
+        self.T_DECAY = 1e-1
 
-    def makePatterns(self):
-        self.eta = scipy.sparse.random(
-            self.N, self.P, density=self.f, data_rvs=lambda n: np.ones((n,))
+    def make_patterns(self):
+        self.patterns = scipy.sparse.random(
+            self.NUM_NEURONS,
+            self.NUM_MEMORIES,
+            density=self.SPARSITY,
+            data_rvs=lambda n: np.ones((n,)),  # makes sparsity definition as in paper
         )
 
     def gain(self, x):
-        x1 = np.where((self.th0 + x) <= 0.0, 0, self.th0 + x)
-        return np.power(x1, self.gamma)
+        # print(x)
+        x1 = np.where((self.GAIN_THRESHOLD + x) <= 0.0, 0, self.GAIN_THRESHOLD + x)
+        return np.power(x1, self.GAIN_EXPONENT)
 
-    def Jr(self, c):
-        r = self.gain(c)
-        pc = self.eta.T.dot(r) - self.f * r.sum()
+    def Jr(self, currents):
+        activations = self.gain(currents)
+        p_currents = (
+            self.patterns.T.dot(activations) - self.SPARSITY * activations.sum()
+        )
         return (
-            (self.eta.dot(pc) - self.f * pc.sum() - r * self.phi) * self.kappa / self.N
+            (
+                self.patterns.dot(p_currents)
+                - self.SPARSITY * p_currents.sum()
+                - activations * self.OSCILLATION
+            )
+            * self.EXCITATION
+            / self.NUM_NEURONS
         )
 
-    def grad(self, c):
+    def grad(self, currents):
         return (
-            -c + self.Jr(c) + np.random.randn(c.shape[0], c.shape[1]) * self.xi0
-        ) / self.tau
+            -currents
+            + self.Jr(currents)
+            + np.random.randn(currents.shape[0], currents.shape[1]) * self.NOISE_VAR
+        ) / self.T_DECAY
+
 
 #%%
 
 np.random.seed(12354)
 mod = Model()
-mod.N = 10000
-mod.P = 16
-mod.f = 0.1
-# mod.eta*=0
-mod.xi0 = 0.5
-mod.kappa = 1000
[…]
