### 2020-05-28 #10 @c-torre
Hi all, I hope everything is well. We would like to update on the state of the replication, especially since many things happened since we contacted all authors.

We got replies and clarifications for a few questions we had. They spotted and confirmed that several typos were introduced into the equations that made the replication impossible.

The first author got a way back into the cluster where the original files were and found that the parameters in the cluster simulations were different than the ones in the previously provided script, and also than those reported in the paper.

As of right now, it seems that the first author could finally fully simulate all the figures, which should complete the replication work.
We still have to launch the longer simulations ourselves confirm replication on our side, and then proceed to translate and clean the code, and update the writing accordingly.

The first and last authors expressed their willingness to publish a correction on the original article, especially when this replication paper is published so they can provide a direct link to an implementation, and credit us for helping in the corrections.

We are working on the simulations now and will be back with hopefully good news
