### 2020-05-28 #10 @oliviaguest
@c-torre: Thank you so much for the update and well done on all the work in figuring this out! I am very pleased to have pushed (lightly) towards you contacting them — this is highly productive for the journal/this publication... and I hope also has been useful for *you* and *your* work as well as, of course, science generally! Great work and I am excited to see the next steps. 👏
