+++ analysismodelmisha.py
@@ -1,0 +1,196 @@
+# works
+#%%
+import glob
+import os
+import pickle
+
+#%% Figure 3
+import matplotlib.pylab as pylab
+import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+# from scipy.io import loadmat
+from scipy.sparse import coo_matrix, csr_matrix
+from tqdm import tqdm
+
+# folder = "you_local_folder_of_choice"
+filename = "sim_"
+
+# files = glob.glob(folder + filename + "_*")
+results_dir = "./results/"
+files_names = os.listdir(results_dir)
+file_paths = sorted([results_dir + file_name for file_name in files_names])
+file_paths = file_paths
+files = file_paths
+print(files)
+
+#%%
+rates_list = []
+similarity_list = []
+for i_file, file_ in tqdm(enumerate(files)):
+    print(i_file)
+    print(file_)
+    target = file_
+    if os.path.getsize(target) > 0:
+        with open(target, "rb") as f:
+            unpickler = pickle.Unpickler(f)
+            # if file is not empty scores will be equal
+            # to the value unpickled
+            rates = unpickler.load()
+    #rates = pickle.load(open(file_, "rb"))
+    rates_list.append(rates["rate_avg"])
+    similarity_list.append(rates["similarity"])
+
+#%%
+# rates_np = np.concatenate(rates_list)
+rates_np = np.stack(rates_list)
+N_trials = rates_np.shape[0]
+rates_npmax = rates_np[
+    :, :, ::100
+]  # This selects out the maximum values of the rates oscillation which occur every 100 steps
+# rates_npmax = rates_np[:, ::100]
+#%%
+rates_npmax[:, :, 0] = rates_np[
+    :, :, 1
+]  # This is because the max of the first item is not at position 0 but at position 1
+
+# rates_npmax[:, 0] = rates_np[:, 1]
+
+sequence = np.argmax(rates_npmax, axis=1)
+T_max = sequence.shape[1]
+# T_max = sequence.shape[0]
+sequence_max = np.max(rates_npmax, axis=1)
+sequence = sequence * (sequence_max > 15)
+
+similarities = np.stack(similarity_list)
+# similarities = np.concatenate(similarity_list)
+#%%
+
+df = pd.DataFrame()
+for i_trial in tqdm(range(N_trials)):
+    items, times = np.unique(sequence[i_trial], return_index=True)
+    times = times[items != 0]
+    times = np.sort(times)
+    items = sequence[i_trial][times]
+    irts = np.insert(np.diff([times]), 0, 0)
+    sizes = similarities[i_trial][items, items]
+    ranksall = np.array([np.argsort(x).argsort() for x in similarities[i_trial]])
+    intersections = np.insert(similarities[i_trial][items[:-1], items[1:]], 0, 0)
+    ranks = np.insert(ranksall[items[:-1], items[1:]], 0, 0)
+    dicttrial = {
+        "items": items,
+        "times": times,
+        "trial": i_trial,
+        "irts": irts,
+        "sizes": sizes,
+        "intersections": intersections,
+        "ranks": ranks,
+    }
+    dftrial = pd.DataFrame.from_dict(dicttrial)
+    df = df.append(dftrial)
+
+#%%
+df = df.reset_index().rename(columns={"index": "position"})
+N_recalled = (
+    df.groupby("trial")
+    .agg({"position": np.max})
[…]
