+++ main.py
@@ -1,336 +1,265 @@
-"""
-<[Re] Recanatesi (2015). Neural Network Model of Memory Retrieval>
-Copyright (C) <2019>  <de la Torre-Ortiz C, Nioche A>
-
-This program is free software: you can redistribute it andtimeor modify
-it unt_stepe terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <https://www.gnu.org/licenses/>.
-"""
-
+# %%
 import os
-import pickle
-from multiprocessing import cpu_count
-from multiprocessing.pool import Pool
+import pickle  # to be replaces with .npy files
 
+from scipy.stats import norm
+import matplotlib.pyplot as plt
+import matplotlib.style as mpstyle
 import numpy as np
-from tqdm import tqdm
 
-import tools.plots as plot
-import tools.recall_performance
-import tools.sine_wave
-
-FIG_DIR = "fig"
-
-os.makedirs(FIG_DIR, exist_ok=True)
-
-np.seterr(all="raise")
+RESULTS_DIR = "results"
+# To use when refactoring without pickle dictionary
+# ACTIVITIES_DIR = os.path.join(RESULTS_DIR, "activities")
+# SIMILARITIES_DIR = os.path.join(RESULTS_DIR, "similarities")
 
 
-# === Parameters ===
-# Architecture
-NUM_NEURONS = 10 ** 5
-NUM_MEMORIES = 16
+try:
+    job_env = int(os.getenv("SLURM_ARRAY_TASK_ID")) # Changes seed per cluster simulation
+except:
+    job_env = 0 # Default seed for non-cluster use
+np.random.seed(job_env)
+
+mpstyle.use("fast")  # Faster plots for prototyping
+
+NUM_NEURONS = 10 ** 5  # "N"
+NUM_MEMORIES = 16  # "P"
 # Activation
-T_DECAY = 0.01
-# Gain
-GAIN_THRESHOLD = 0
-GAIN_EXP = 2 / 5
-# Hebbian rule
-EXCITATION = 13_000
-SPARSITY = 0.1
-# Inhibition
-SIN_MIN = 0.7
-SIN_MAX = 1.06
-T_OSCILLATION = 1
-PHASE_SHIFT = 0.75
-# Short term association
-CONT_FORTH = 1500
-CONT_BACK = 400
+T_DECAY = 0.01  # "tau"
 # Time
-T_CYCLES = 8.0  # 14
-T_STEP = 0.001
+T_STEP = 0.001  # "dt"
+T_TOT = 450 # "T"; Half time iterations for prototyping (default: 450)
+T_SIMULATED = int(T_TOT // T_STEP)
+# Hebbian rule
+EXCITATION = 12_500  # "J1"
+SPARSITY = 0.1  # "f0"
+# Gain
+GAIN_THRESHOLD = 0  # set as *(x > 0) in gain function in matlab
+GAIN_EXP = 1 / 3  # set in gain function in matlab
+# Inhibition
+SIN_MIN = 0.4 * EXCITATION  # "J_0m"
+SIN_MAX = 1.2 * EXCITATION  # "J_0M"
 # Noise
-NOISE_VAR = 65
-# Initialization
-INIT_RATE = 1
-FIRST_MEMORY = 7
-# Recall
-RECALL_THRESHOLD = 15
[…]
