### 2020-08-04 9a09a47
scripts and params
