+++ tools/__init__.py
@@ -1,10 +1,0 @@
-"""
-Copyright (C) <2019>  <de la Torre-Ortiz C, Nioche A>
-See the 'LICENSE' file for details.
-
-Network model utilities.
-
-Package contains modules:
-    plots
-    sine_wave
-"""
