### 2020-08-04 #10 @c-torre
Dear all,
Please find an updated version of the code as in the original issue. The PDF link in the original issue link will take you to our previous manuscript submission if needed for comparison. The new PDF can be found at: https://github.com/c-torre/replication-recanatesi-2015/blob/full-replication/re-neural-network-model-memory-retrieval.pdf

We are thankful for the reviewer's comments as they have drastically improved the quality of our work. We also thank all the discussion, as it led the replication work in the appropriate direction. We would encourage to revisit both the manuscript and code, as very major changes have taken place. Here we address the review comments:

#### Review 1

* Why is the sparsity parameter different? Shouldn't it be 0.1?

A typo was found and corrected the parameter to its correct value of 0.1

* Where is the recall activation threshold parameter defined?

We followed the approach of the original interpretation now, were recalls are consistently observed at every inhibition minima, where rates always go above this threshold value. In the implementation this means that recalls are calculated not based on a threshold, but on the time iteration at which the peak maximum occurs.

* Why is there no attempt to plot something similar to figures 3 to 6?

We have corrected this major issue and now provide missing and extra figures.

#### Review 2

* Major Comment: What accounts for the difference between the results with the original and reproduction variable? As a work aimed at reproduction, presenting differences between dynamics with the original parameters and the scaled reproduction variables would be useful. In some cases parameters need to be changed by N (10^5). Presentation of the differences would be helpful to understand this.

We have been collaborating with the original authors as we discovered errors in the original article. Normalization of parameters was indeed one of the main issues with the original paper. We have addressed this in detail in the new manuscript, as it is the core of the major changes since our original review request.

* Minor Comments:
After running the network for several different seeds, I found that there were several periods where the same memory was recalled in sequence. The original article presents a simulation with up to three repetitions while the provided code often produces repetitions of 6 or longer, even 14 in one case. Is this a meaningful difference resulting from the difference in parameters?

A new section has been added with the new figures, which address more in detail changes to the seed, as 10000 networks with different seeds are simulated to produce the recall analysis figures and conclusions. We agree that Figure 2 can be seed or parameter-dependent, and that many recalls in a row are possible with seed changes as seen in later figures.

In the time you may take to have a look at our replies, we intend to continue collaborating in the correction of the original article, as the original authors expressed wishes to continue working together on this issue and hope encourage future citations to also include our work.

Thank you for all your patience and help,
Best,
Carlos
