### 2020-09-09 #10 @oliviaguest
@c-torre OK, amazing!

Can you please update the metadata.md file with the details of each reviewer, editor, etc., please? Acceptance date is today and also see: https://github.com/ReScience/ReScience/issues/48#issuecomment-689776914! 🥳
