### 2020-09-10 #10 @c-torre
Hi all,

This is very good news. However, we went an extra step to ensure that everything is absolutely correct this time. Thanks to the original authors we have spotted the following mistakes in our current submission:

1. Equation 2 should have the same time scaling as equation 10, adding sqrt(dt) for the noise term.
1. Equation 4 has a mistake in the constant part.
1. Equation 6 has a mistake in the contiguity terms.
1. A parenthesis is lacking in the manuscript text.
1. A parameter value was not updated in the new manuscript text.
1. Parameter table has some issues in the consistency of scaling.
1. Code should be re-run to ensure that is bug-free and replication holds after this tweaking.

While the replication is in a very good place now, as opposed to out first submission, I believe it's in the community's best interest to address these issues before proceeding to final publication.
