### 2020-09-10 #10 @oliviaguest
Sure, @c-torre! I'd assumed you had fixed these. Let me know when it's good to go.
