+++ main.py
@@ -1,3 +1,4 @@
+#%%
 """
 <[Re] Recanatesi (2015). Neural Network Model of Memory Retrieval>
 Copyright (C) <2020>  <de la Torre-Ortiz C, Nioche A>
@@ -21,14 +22,13 @@ import os
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-
 from tqdm import tqdm
 
 import paths
 
 PARAMETERS_DIR = paths.PARAMETERS_DIR
-PATTERNS_DIR = None  # paths. ...
-RECALLS_DIR = None  # paths. ...
+PATTERNS_DIR = paths.PATTERNS_SEEDS_DIR  # paths. ...
+RECALLS_DIR = paths.RECALLS_SEEDS_DIR  # paths. ...
 
 assert (
     PATTERNS_DIR is not None and RECALLS_DIR is not None
@@ -45,28 +45,28 @@ np.random.seed(job_id)
 
 PARAMETERS_DF = pd.read_csv(os.path.join(PARAMETERS_DIR, "simulation.csv"), index_col=0)
 
-NUM_NEURONS = PARAMETERS_DF.loc["num_neurons"]
-NUM_MEMORIES = PARAMETERS_DF.loc["num_memories"]
+NUM_NEURONS = int(PARAMETERS_DF.loc["num_neurons"].array[0])
+NUM_MEMORIES = int(PARAMETERS_DF.loc["num_memories"].array[0])
 # Activation
-T_DECAY = PARAMETERS_DF.loc["t_decay"]
+T_DECAY = PARAMETERS_DF.loc["t_decay"].array[0]
 # Time
-T_STEP = PARAMETERS_DF.loc["t_step"]
-T_TOT = PARAMETERS_DF.loc["t_tot"]
+T_STEP = PARAMETERS_DF.loc["t_step"].array[0]
+T_TOT = 10  # PARAMETERS_DF.loc["t_tot"].array[0] !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 T_SIMULATED = int(T_TOT // T_STEP)
 # Hebbian rule
-EXCITATION = PARAMETERS_DF.loc["excitation"]
-SPARSITY = PARAMETERS_DF.loc["sparsity"]
+EXCITATION = PARAMETERS_DF.loc["excitation"].array[0]
+SPARSITY = PARAMETERS_DF.loc["sparsity"].array[0]
 # Gain
-GAIN_THRESHOLD = PARAMETERS_DF.loc["gain_threshold"]
-GAIN_EXP = PARAMETERS_DF.loc["gain_exp"]
+GAIN_THRESHOLD = PARAMETERS_DF.loc["gain_threshold"].array[0]
+GAIN_EXP = PARAMETERS_DF.loc["gain_exp"].array[0]
 # Inhibition
-SIN_MIN = PARAMETERS_DF.loc["sin_min"] * EXCITATION
-SIN_MAX = PARAMETERS_DF.loc["sin_max"] * EXCITATION
+SIN_MIN = PARAMETERS_DF.loc["sin_min"].array[0] * EXCITATION
+SIN_MAX = PARAMETERS_DF.loc["sin_max"].array[0] * EXCITATION
 # Noise
-NOISE_VAR = PARAMETERS_DF.loc["noise_var"]
+NOISE_VAR = PARAMETERS_DF.loc["noise_var"].array[0]
 # Forward and backward contiguity
-CONT_FORTH = PARAMETERS_DF.loc["cont_forth"] / NUM_NEURONS
-CONT_BACK = PARAMETERS_DF.loc["cont_back"] / NUM_NEURONS
[…]
