### 2020-11-12 3af1f1d
debug plots
