+++ plot.py
@@ -7,19 +7,16 @@ Plotting functions.
 """
 
 import os
-from typing import Hashable, Iterable
 
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-from scipy.io import loadmat
-from scipy.sparse import coo_matrix, csr_matrix
+from scipy.sparse import csr_matrix
 
 import file_loader
 import paths
 import recall_performance
 
-#
 # def plot_parameter_sweep(
 #     parameter: str,
 #     sims_path: str,
@@ -34,10 +31,14 @@ import recall_performance
 #     """Plots the parameter sweeps"""
 #
 #     def plot(
-#         df: pd.DataFrame, xlabel: str, ylabel: str, fig_file_path: str, fig_name: str,
+#         df: pd.DataFrame,
+#         xlabel: str,
+#         ylabel: str,
+#         fig_file_path: str,
+#         fig_name: str,
 #     ) -> None:
 #
-#         plt.plot(dft["param"], dft["N_recalled"], "-o", linewidth="0.5")
+#         plt.plot(df["param"], df["num_memories_recalled"], "-o", linewidth="0.5")
 #         plt.xlabel(xlabel)
 #         plt.ylabel(ylabel)
 #         plt.yticks(range(16))
@@ -49,7 +50,10 @@ import recall_performance
 #
 #     # File names will be parsed to get parameters
 #     file_names = sorted(
-#         file_loader.get_file_names_checked({patterns_path, recalls_path}, name_filter,)
+#         file_loader.get_file_names_checked(
+#             {patterns_path, recalls_path},
+#             name_filter,
+#         )
 #     )
 #
 #     file_paths_patterns = sorted(file_loader.get_file_paths(patterns_path, name_filter))
@@ -63,52 +67,22 @@ import recall_performance
 #     )
 #
 #     recalls_sequence = recall_performance.load_sequences(file_paths_recalls)
+#     print("I'M GOING TO PRINT RECALL SEQUENCES HEHEHEH")
+#     print(recalls_sequence)
+#     print("I DID IT ALREADY MUAHAHAHA")
 #
-#     similarities = recall_performance.make_similarity(file_paths_patterns, num_trials)
+#     similarities = recall_performance.make_similarity(
+#         file_paths_patterns
+#     )  # , num_trials)
 #
 #     dft = recall_performance.get_main_df(
-#         num_trials, parameters_used[parameter], recalls_sequence, similarities,
+#         num_trials,
+#         parameters_used[parameter],
+#         recalls_sequence,
+#         similarities,
 #     )
 #
-#     plot(dft, xlabel, ylabel, paths.FIGURES_DIR, fig_name)
-#
-#
-# plot_parameter_sweep(
-#     parameter="noise_var",
-#     sims_path=paths.SIMS_NOISE_STD_DIR,
-#     patterns_path=paths.PATTERNS_NOISE_STD_DIR,
-#     recalls_path=paths.RECALLS_NOISE_STD_DIR,
-#     name_filter=".npy",
-#     xlabel="Noise standard deviation",
-#     ylabel="Average words recalled",
-#     fig_file_path=paths.FIGURES_DIR,
-#     fig_name="noise_var.pdf",
-# )
-#
-# plot_parameter_sweep(
-#     parameter="cont_forth",
-#     sims_path=paths.SIMS_CONT_FORTH_DIR,
-#     patterns_path=paths.PATTERNS_CONT_FORTH_DIR,
-#     recalls_path=paths.RECALLS_CONT_FORTH_DIR,
-#     name_filter=".npy",
-#     xlabel="Forward contiguity",
-#     ylabel="Average words recalled",
[…]
