+++ recall_performance.py
@@ -6,18 +6,16 @@ Utils for plotting network recalls analysis.
 """
 
 import os
-from typing import Hashable, Iterable
+import re
+from typing import Iterable
 
-import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-from scipy.io import loadmat
-from scipy.sparse import coo_matrix, csr_matrix
 from tqdm import tqdm
 
 
 def get_used_parameters(
-    file_names: Hashable,
+    file_names: Iterable,
     sims_path: str,
     num_trials: int,
     separator="-",
@@ -25,27 +23,26 @@ def get_used_parameters(
     """Gets and saves the set of parameters of the simulations"""
 
     def get_changing_parameter(file_name: str, position: int, separator="-") -> int:
-        """Parse name to get parameters"""
+        """Parse file name to get parameters"""
 
-        return file_name.split(separator)[position]
+        root, _ = os.path.splitext(file_name)
+        selected_param = root.split(separator)[position]
+        match = re.search("[0-9]+$", selected_param)  # Match any last number(s)
+
+        return int(match.group())
+
+    changing_parameters = "seed", "cont_forth", "noise_var"
 
     parameters_used = pd.DataFrame(
         index=range(num_trials),
-        columns=["seed", "cont_forth", "noise_var"],
+        columns=changing_parameters,
     )
 
-    parameters_used["seed"] = tuple(
-        int(get_changing_parameter(file_name, position=0, separator=separator)[1:])
-        for file_name in file_names
-    )
-    parameters_used["cont_forth"] = tuple(
-        int(get_changing_parameter(file_name, position=1)[2:])
-        for file_name in file_names
-    )
-    parameters_used["noise_var"] = tuple(
-        int(get_changing_parameter(file_name, position=2)[1:-4])
-        for file_name in file_names
-    )
+    for idx, parameter_name in enumerate(changing_parameters):
+        parameters_used[parameter_name] = tuple(
+            (get_changing_parameter(file_name, position=idx, separator=separator))
+            for file_name in file_names
+        )
 
     parameters_used.to_csv(os.path.join(sims_path, "parameters.csv"))
     return parameters_used
@@ -57,7 +54,7 @@ def load_sequences(file_paths: Iterable) -> None:
     print("Loading recall sequences...")
     sequences_shapes = pd.DataFrame(columns=("Recall sequence", "Length"))
 
-    for file_idx, file_path in enumerate(file_paths):
+    for file_path in tqdm(file_paths):
         recall_sequence = np.load(file_path)
         new_row = {"Recall sequence": recall_sequence, "Length": len(recall_sequence)}
         sequences_shapes = sequences_shapes.append(new_row, ignore_index=True)
@@ -73,17 +70,20 @@ def load_sequences(file_paths: Iterable) -> None:
 
 
 def make_similarity(
-    similarity_paths: str,
-    # num_trials: int,
+    similarity_paths: Iterable,
 ) -> np.ndarray:
     """Get pattern similarities from saved ndarrays"""
 
     print("Loading pattern similarities...")
-    similarities = np.stack(
-        [np.load(similarity_path) for similarity_path in sorted(similarity_paths)]
-    )
+    similarities = []
+    for similarity_path in tqdm(sorted(similarity_paths)):
+        similarities.append(np.load(similarity_path))
+
+    # similarities = np.stack(
+    #     [np.load(similarity_path) for similarity_path in sorted(similarity_paths)]
+    # )
     print("Done!")
-    return similarities
+    return np.stack(similarities)
 
 
 def get_main_df(
@@ -95,50 +95,49 @@ def get_main_df(
     """Make dataframe with all recall metrics"""
 
[…]
