### 2020-12-04 ff6673f
cleaning and fixes
