+++ plot.py
@@ -1,4 +1,3 @@
-#%%
 """
 Copyright (C) <2019-2020>  <de la Torre-Ortiz C, Nioche A>
 See the 'LICENSE' file for details.
@@ -11,105 +10,52 @@ import os
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
+import paths
 from scipy.sparse import csr_matrix
 
-import file_loader
-import paths
-import recall_performance
-
-# def plot_parameter_sweep(
-#     parameter: str,
-#     sims_path: str,
-#     patterns_path: str,
-#     recalls_path: str,
-#     name_filter: str,
-#     xlabel: str,
-#     ylabel: str,
-#     fig_file_path: str,
-#     fig_name: str,
-# ) -> None:
-#     """Plots the parameter sweeps"""
-#
-#     def plot(
-#         df: pd.DataFrame,
-#         xlabel: str,
-#         ylabel: str,
-#         fig_file_path: str,
-#         fig_name: str,
-#     ) -> None:
-#
-#         plt.plot(df["param"], df["num_memories_recalled"], "-o", linewidth="0.5")
-#         plt.xlabel(xlabel)
-#         plt.ylabel(ylabel)
-#         plt.yticks(range(16))
-#         plt.tight_layout()
-#         plt.savefig(os.path.join(fig_file_path, fig_name))
-#         plt.close("all")
-#
-#     print(f"Preparing {fig_name} plot... ...")
-#
-#     # File names will be parsed to get parameters
-#     file_names = sorted(
-#         file_loader.get_file_names_checked(
-#             {patterns_path, recalls_path},
-#             name_filter,
-#         )
-#     )
-#
-#     file_paths_patterns = sorted(file_loader.get_file_paths(patterns_path, name_filter))
-#     file_paths_recalls = sorted(file_loader.get_file_paths(recalls_path, name_filter))
-#
-#     assert len(file_names) == len(file_paths_patterns) == len(file_paths_recalls)
-#     num_trials = len(file_paths_patterns)
-#
-#     parameters_used = recall_performance.get_used_parameters(
-#         file_names, sims_path, num_trials
-#     )
-#
-#     recalls_sequence = recall_performance.load_sequences(file_paths_recalls)
-#     print("I'M GOING TO PRINT RECALL SEQUENCES HEHEHEH")
-#     print(recalls_sequence)
-#     print("I DID IT ALREADY MUAHAHAHA")
-#
-#     similarities = recall_performance.make_similarity(
-#         file_paths_patterns
-#     )  # , num_trials)
-#
-#     dft = recall_performance.get_main_df(
-#         num_trials,
-#         parameters_used[parameter],
-#         recalls_sequence,
-#         similarities,
-#     )
-#
-#     plot(dft, xlabel, ylabel, fig_file_path, fig_name)
+import utils.file_loading as loading
+from utils import recall_performance
 
 
-def plot_analysis(
+def plot_parameter_sweep(
+    parameter: str,
     sims_path: str,
     patterns_path: str,
     recalls_path: str,
     name_filter: str,
-    fig_file_path: str,
+    xlabel: str,
+    ylabel: str,
+    fig_name: str,
 ) -> None:
-    """
-    Plot all figures but for the parameter sweep.
-    Takes advantage of only computing the dataframe once
-    therefore it is a longer method.
-    """
+    """Plots the parameter sweeps"""
 
-    print("Preparing temporal properties plots... ...")
+    def plot(
+        df: pd.DataFrame,
+        xlabel: str,
+        ylabel: str,
[…]
