+++ main.py
@@ -1,291 +1,0 @@
-#%%
-"""
-<[Re] Recanatesi (2015). Neural Network Model of Memory Retrieval>
-Copyright (C) <2020>  <de la Torre-Ortiz C, Nioche A>
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program. If not, see <https://www.gnu.org/licenses/>.
-"""
-
-import os
-
-import matplotlib.pyplot as plt
-import numpy as np
-import pandas as pd
-from tqdm import tqdm
-
-import paths
-
-PARAMETERS_DIR = paths.PARAMETERS_DIR
-PATTERNS_DIR = paths.PATTERNS_SEEDS_DIR  # paths. ...
-RECALLS_DIR = paths.RECALLS_SEEDS_DIR  # paths. ...
-
-assert (
-    PATTERNS_DIR is not None and RECALLS_DIR is not None
-), "Please choose a saving directory"
-
-try:
-    job_id = int(
-        os.getenv("SLURM_ARRAY_TASK_ID")
-    )  # Changes seed per cluster simulation
-except:
-    job_id = 0  # Default seed for non-cluster use
-
-np.random.seed(job_id)
-
-PARAMETERS_DF = pd.read_csv(os.path.join(PARAMETERS_DIR, "simulation.csv"), index_col=0)
-
-NUM_NEURONS = int(PARAMETERS_DF.loc["num_neurons"].array[0])
-NUM_MEMORIES = int(PARAMETERS_DF.loc["num_memories"].array[0])
-# Activation
-T_DECAY = PARAMETERS_DF.loc["t_decay"].array[0]
-# Time
-T_STEP = PARAMETERS_DF.loc["t_step"].array[0]
-T_TOT = PARAMETERS_DF.loc["t_tot"].array[0]
-T_SIMULATED = int(T_TOT // T_STEP)
-# Hebbian rule
-EXCITATION = PARAMETERS_DF.loc["excitation"].array[0]
-SPARSITY = PARAMETERS_DF.loc["sparsity"].array[0]
-# Gain
-GAIN_THRESHOLD = PARAMETERS_DF.loc["gain_threshold"].array[0]
-GAIN_EXP = PARAMETERS_DF.loc["gain_exp"].array[0]
-# Inhibition
-SIN_MIN = PARAMETERS_DF.loc["sin_min"].array[0] * EXCITATION
-SIN_MAX = PARAMETERS_DF.loc["sin_max"].array[0] * EXCITATION
-# Noise
-NOISE_VAR = PARAMETERS_DF.loc["noise_var"].array[0]
-# Forward and backward contiguity
-CONT_FORTH = PARAMETERS_DF.loc["cont_forth"].array[0] / NUM_NEURONS
-CONT_BACK = PARAMETERS_DF.loc["cont_back"].array[0] / NUM_NEURONS
-# For parameter sweeps
-# CONT_FORTH = get_simulation_range_param("cont_forth", job_id, 100)
-# CONT_FORTH = get_simulation_range_param("cont_forth_low", job_id, 100)
-# NOISE_VAR = get_simulation_range_param("noise_var", job_id, 100)
-
-
-def get_simulation_range_param(sim_mode: str, job_id: int, num_points: int) -> float:
[…]
