+++ run.sh
@@ -1,8 +1,8 @@
-#!/bin/bash
+#!/bin/sh
 
-truncate --size 0 out/log.out
+mkdir -p debug
+truncate --size 0 debug/log.out
 echo -n $@
 PID=$(sbatch --requeue --parsable "$@" | tail -n1)
 echo "	$PID"
 echo $PID > .last.pid
-#tail -f out/test.out
