+++ settings/paths.py
@@ -7,7 +7,7 @@ Directory management.
 
 import os
 
-BASE_DIR = os.path.dirname(os.path.abspath(__file__))
+BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
 #
 PARAMETERS_DIR = os.path.join(BASE_DIR, "parameters")
