### 2020-12-07 0987349
cleaned figures
