+++ simulation.py
@@ -74,7 +74,7 @@ NOISE_VAR = PARAMETERS_DF.loc["noise_var"].array[0]
 CONT_FORTH = PARAMETERS_DF.loc["cont_forth"].array[0] / NUM_NEURONS
 CONT_BACK = PARAMETERS_DF.loc["cont_back"].array[0] / NUM_NEURONS
 # For parameter sweeps (uncomment to select)
-# CONT_FORTH = sim.get_simulation_range_param("cont_forth", JOB_ID, 100) / NUM_NEURONS
+CONT_FORTH = sim.get_simulation_range_param("cont_forth", JOB_ID, 100) / NUM_NEURONS
 # CONT_FORTH = sim.get_simulation_range_param("cont_forth_low", JOB_ID, 100) / NUM_NEURONS
 # NOISE_VAR = sim.get_simulation_range_param("noise_var", JOB_ID, 100)
 
