+++ plot.py
@@ -6,13 +6,14 @@ Plotting functions.
 """
 
 import os
+import string
 
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-import paths
 from scipy.sparse import csr_matrix
 
+import settings.paths as paths
 import utils.file_loading as loading
 from utils import recall_performance
 
@@ -25,7 +26,7 @@ def plot_parameter_sweep(
     name_filter: str,
     xlabel: str,
     ylabel: str,
-    fig_name: str,
+    file_name: str,
 ) -> None:
     """Plots the parameter sweeps"""
 
@@ -33,7 +34,7 @@ def plot_parameter_sweep(
         df: pd.DataFrame,
         xlabel: str,
         ylabel: str,
-        fig_name: str,
+        file_name: str,
     ) -> None:
 
         plt.plot(df["param"], df["num_memories_recalled"], "-o", linewidth="0.5")
@@ -41,10 +42,10 @@ def plot_parameter_sweep(
         plt.ylabel(ylabel)
         plt.yticks(range(16))
         plt.tight_layout()
-        plt.savefig(os.path.join(paths.FIGURES_DIR, fig_name))
+        plt.savefig(os.path.join(paths.FIGURES_DIR, file_name))
         plt.close("all")
 
-    print(f"Preparing {fig_name} plot... ...")
+    print(f"Preparing {file_name} plot... ...")
 
     # File names will be parsed to get parameters
     file_names = sorted(
@@ -75,9 +76,9 @@ def plot_parameter_sweep(
         similarities,
     )
     # Calculate for the varying parameter
-    metrics_df = metrics_df.groupby("param")
+    metrics_df = metrics_df.groupby("param").apply(np.mean)
 
-    plot(metrics_df, xlabel, ylabel, paths.FIGURES_DIR, fig_name)
+    plot(metrics_df, xlabel, ylabel, file_name)
 
 
 def plot_analysis(name_filter: str) -> None:
@@ -124,9 +125,11 @@ def plot_analysis(name_filter: str) -> None:
 
     def plot_time_words_recalled(
         recalls_df: pd.DataFrame,
-        fig_name: str,
+        panel_num: int,
+        file_name: str,
     ) -> None:
 
+        print("Plotting Average Words Recalled...")
         mid_memory = 7
         data_sparse = np.ones_like(recalls_df.trial.values.astype(int))
         recalls_bin = csr_matrix(
@@ -142,36 +145,61 @@ def plot_analysis(name_filter: str) -> None:
         recalls_mean = np.mean(recalls_bincum, axis=0) * mid_memory
         recalls_mean = recalls_mean * mid_memory / 50
 
-        plt.plot(recalls_mean)
-        plt.xlabel("Time steps")
-        plt.ylabel("Average words recalled")
-        plt.tight_layout()
-        plt.savefig(os.path.join(paths.FIGURES_DIR, fig_name))
-        plt.close("all")
+        fig, axis = plt.subplots()
+        axis.plot(recalls_mean)
+        axis.set_title("Average Words Recalled")
+        axis.set_xlabel("Time steps")
+        axis.set_ylabel("Average words recalled")
 
-    plot_time_words_recalled(metrics_df, paths.FIGURES_DIR, "time_words_recalled.pdf")
+        axis.text(
+            -0.2,
+            1.2,
+            string.ascii_uppercase[panel_num],
+            transform=axis.transAxes,
+            size=20,
+            weight="bold",
+        )
+
+        plt.tight_layout()
+        plt.savefig(os.path.join(paths.FIGURES_DIR, file_name))
+        plt.close("all")
+        print("Done!")
+
[…]
