+++ run.sh
@@ -1,8 +1,9 @@
 #!/bin/sh
 
-mkdir -p debug
-truncate --size 0 debug/log.out
+logs_dir="logs"
+mkdir -p $logs_dir
+truncate --size 0 $logs_dir/log.out
 echo -n $@
 PID=$(sbatch --requeue --parsable "$@" | tail -n1)
 echo "	$PID"
-echo $PID > .last.pid
+echo $PID > .last_pid
