+++ metadata.tex
@@ -1,44 +1,0 @@
-% DO NOT EDIT - automatically generated from metadata.yaml
-
-\def \codeURL{https://github.com/c-torre/replication-recanatesi-2015}
-\def \codeDOI{}
-\def \dataURL{}
-\def \dataDOI{}
-\def \editorNAME{}
-\def \editorORCID{}
-\def \reviewerINAME{}
-\def \reviewerIORCID{}
-\def \reviewerIINAME{}
-\def \reviewerIIORCID{}
-\def \dateRECEIVED{08 November 2019}
-\def \dateACCEPTED{}
-\def \datePUBLISHED{}
-\def \articleTITLE{[Re] Neural Network Model of Memory Retrieval}
-\def \articleTYPE{Replication}
-\def \articleDOMAIN{Computational Neuroscience}
-\def \articleBIBLIOGRAPHY{bibliography.bib}
-\def \articleYEAR{2019}
-\def \reviewURL{}
-\def \articleABSTRACT{}
-\def \replicationCITE{}
-\def \replicationBIB{}
-\def \replicationURL{}
-\def \replicationDOI{}
-\def \contactNAME{Carlos de la Torre-Ortiz}
-\def \contactEMAIL{ctorre@mailbox.org}
-\def \articleKEYWORDS{rescience c, rescience x}
-\def \journalNAME{ReScience C}
-\def \journalVOLUME{5}
-\def \journalISSUE{3}
-\def \articleNUMBER{}
-\def \articleDOI{}
-\def \authorsFULL{Carlos de la Torre-Ortiz and Aurélien Nioche}
-\def \authorsABBRV{C. de la Torre-Ortiz and A. Nioche}
-\def \authorsSHORT{de la Torre-Ortiz and Nioche}
-\title{\articleTITLE}
-\date{}
-\author[1,\orcid{0000-0002-7457-5216}]{Carlos de la Torre-Ortiz}
-\author[2,\orcid{0000-0002-0567-2637}]{Aurélien Nioche}
-\affil[1]{University of Helsinki, Department of Computer Science, Helsinki, Finland}
-\affil[2]{Aalto University, School of Electrical Engineering, Department Communications and Networking, Espoo, Finland}
-
