### 2020-12-28 #10 @c-torre
Hi! Manuscript is ready and the code has a Zenodo DOI now. What's next?

[2020_rescience_neural_network_model_of_memory_retrieval.pdf](https://github.com/ReScience/submissions/files/5748742/2020_rescience_neural_network_model_of_memory_retrieval.3.pdf)
