### 2020-12-31 #10 @rougier
We'll publish the paper next week most probably. Don't hesitate to remind us.
