### 2021-01-18 #10 @c-torre
Added article source with metadata.yaml at [c-torre/replication-recanatesi-2015-article](https://github.com/c-torre/replication-recanatesi-2015-article)
