### 2021-01-22 #10 @c-torre
Updated template, corrected latex errors and added code to software heritage ([c-torre/replication-recanatesi-2015-article](https://github.com/c-torre/replication-recanatesi-2015-article))
