### 2021-01-25 #10 @rougier
Thanks, I'll try to publish it today. For the swh link, can you send the swid (instead of the url) ? You should have received it when you saved the repo. The article template will complete the full url based on this id.
