### 2021-01-25 #10 @rougier
Here is the sanboxed version (not the real one): https://sandbox.zenodo.org/record/719803
Can you check everything looks good ?
