### 2021-01-25 #10 @c-torre
Sandboxed version looks good.

For the SWID, the only thing I seem to find is a "permalink" hidden in a SWH side bar:

```
swh:1:dir:0b541aeb4707ecedfbbcdd85adfd0100d748cc03;
origin=https://github.com/c-torre/replication-recanatesi-2015/;
visit=swh:1:snp:0fd6c99fcc42679041cee0c384bfd4a0ffa0a50e;
anchor=swh:1:rev:59a3c88f5e7af41d4bf748ac4c41e03bdb2722da
```

https://archive.softwareheritage.org/browse/origin/directory/?origin_url=https://github.com/c-torre/replication-recanatesi-2015/

Is it any of these?
