### 2021-01-25 #10 @rougier
Your article is published at https://zenodo.org/record/4461767. It will appear soon on the ReScience website as well.
Congratulations !
